% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AFSCforeign_hauls}
\alias{AFSCforeign_hauls}
\title{Example data for AFSCforeign_hauls}
\format{
A data frame.
}
\usage{
AFSCforeign_hauls
}
\description{
Example data for AFSCforeign_hauls
}
\keyword{datasets}
