% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_length_comparison.fn.R
\name{age_representativeness_plot}
\alias{age_representativeness_plot}
\title{plot age representativeness}
\usage{
age_representativeness_plot(
  bio.WCGBTS,
  xlim = c(0, 120),
  ylim = c(0, 0.049),
  file = NULL
)
}
\arguments{
\item{bio.WCGBTS}{data frame}

\item{xlim}{x limits for plot, defaults to c(0,120)}

\item{ylim}{y limits for plot, defaults to (0, 0.049)}

\item{file}{Defaults to NULL (plot made, but not saved to file). Can alternatively be a string filename
by year and sex}
}
\description{
plot age representativeness
}
