% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{availablecells}
\alias{availablecells}
\title{Available grid cell locations for the survey}
\format{
A data frame with fourteen columns,
Cent.ID, Depth.Range, Lat.34.5, Hectares, Cent.Lat, Cent.Long,
SE.LON, SW.LAT, NW.LAT, NE.LON, NE.LAT, SE.LON, SE.LAT
}
\usage{
availablecells
}
\description{
Available grid cell locations that were used to choose the
2018 survey locations for the WCGBTS. Cells are a truncated
set compared to the full set because of closures to trawlable
areas such as what would be closed because of a new
management measure.
In the future, this table should be available from the
data warehouse itself rather than a static snapshot from
a single year.
}
\author{
Eric Ward
}
\keyword{datasets}
