% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_nwfsc_combo}
\alias{bio_nwfsc_combo}
\title{Example data for NWFSC bio data}
\format{
A data frame.
}
\usage{
bio_nwfsc_combo
}
\description{
Example data for NWFSC bio data
}
\keyword{datasets}
