% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_nwfsc_combo}
\alias{catch_nwfsc_combo}
\title{Example data for NWFSC catch data}
\format{
A data frame.
}
\usage{
catch_nwfsc_combo
}
\description{
Example data for NWFSC catch data
}
\keyword{datasets}
