% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strata.R
\name{check_strata}
\alias{check_strata}
\title{Calculate the number of observations by year and strata}
\usage{
check_strata(
  data,
  strata,
  dir = NULL,
  printfolder = lifecycle::deprecated(),
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame of catch data created by \code{\link[=pull_catch]{pull_catch()}}.}

\item{strata}{A data frame that defines the strata and provides the
calculated areas for each strata returned from \code{\link[=create_strata]{create_strata()}}.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{printfolder}{Deprecated with {nwfscSurvey} 2.8.0 to give users greater
control on where to save files. A string that will be appended to \code{dir},
creating a folder where the output will be saved.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
A matrix containing the total number of tows and positive tows
cross-tabulated by year and strata.
}
\description{
Provides a summary of total and positive tows across different strata and years
to help evaluate the suitability of the selected stratification.
}
\details{
Calculates and returns the total number of tows and
positive tows conducted in each strata by year. The
selected strata are used to expand the length and
marginal age compositions and to calculate a design-based
index using \code{\link[=get_design_based]{get_design_based()}}.

In earlier versions of the code, more than one positive observation
within each strata was required to calculate a design-based index.
The current \code{\link[=get_design_based]{get_design_based()}} function is more robust and returns
zeros in strata-year combinations with no observations. However,
reviewing the frequency of tows and positive tows is recommended
to ensure that the selected strata are reasonable (e.g., avoiding
limited observations in large areas).
}
\seealso{
Other helper functions: 
\code{\link{codify_sex}()}
}
\author{
Chantel Wetzel
}
\concept{helper functions}
