% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tows.R
\name{combine_tows}
\alias{combine_tows}
\title{Combine data with multiple records for unique tows}
\usage{
combine_tows(data, dir = NULL, single_species = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A data frame of catches obtained by using \code{\link[=pull_catch]{pull_catch()}}.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{single_species}{TRUE/FALSE Logical that indicates whether data should be
grouped by \code{trawl_id} alone (TRUE, e.g., rougheye rockfish and blackspotted rockfish)
or data should be grouped by \verb{common name} and \code{trawl_id} (FALSE).}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Combine catch data by \code{trawl_id} for data pulled using the
pull_catch function. Generally, only a single tow is returned
by species and in those instances this function is not needed.
For example, if data are pulled with all sample_types included,
then there will often be data from the same \code{trawl_id} included
as multiple records. In these instances, this function allows
for the sample data to be combined for each unique \code{trawl_id}.
This also occurs if cryptic species pairs are all returned
(e.g., rougheye rockfish and blackspotted rockfish)
}
\seealso{
Other helper function:
\code{\link[=check_dir]{check_dir()}},
\code{\link[=check_survey]{check_survey()}},
\code{\link[=createMatrix]{createMatrix()}},
\code{\link[=filter_pull]{filter_pull()}}
}
\author{
Chantel Wetzel
}
\concept{helper function}
