% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_colnames}
\alias{convert_colnames}
\title{Utility function used to convert column names}
\usage{
convert_colnames(x)
}
\arguments{
\item{x}{A data frame with named columns}
}
\value{
A data frame with renamed columns
}
\description{
Utility function used to convert column names
}
\details{
Function that converts the new data warehouse
column names to the previous column names
from the trips/api
}
\author{
Chantel Wetzel
}
