% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_hex_string}
\alias{convert_to_hex_string}
\title{Utility function used throughout the package}
\usage{
convert_to_hex_string(x)
}
\arguments{
\item{x}{A string of either common_name or
sci_name}
}
\description{
Utility function used throughout the package
}
\details{
Function that converts a string to a hex string
for common name or scientific name when pulling
data. This function is used within the pull_*
functions that retrieve species specific data
}
\examples{
\dontrun{
common_name <- c("lingcod", "sablefish", "Pacific cod")
convert_to_hex_string(common_name)
}

}
\author{
Kelli Johnson
}
