% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStrataDF.fn.R
\name{CreateStrataDF.fn}
\alias{CreateStrataDF.fn}
\title{Function that is able to create strata data frame and will calculate strata areas using the StrataAreas.fn}
\usage{
CreateStrataDF.fn(names = NA, depths, lats)
}
\arguments{
\item{names}{vector of names for strata, if not specified the default with be A, B, ...}

\item{lats}{vector of latitude splits for each strata}

\item{depth}{vector depth splits for each strata}

\item{SA3}{A dataframe of min and max latitude and min and max depth with area for each defined square.  See Details.}
}
\value{
Returns the dataframe input with a column called area.
}
\description{
Example function call
createStratDF.fn (names = c('shallow', 'deep'), depths = c(183, 300, 500), lat = c(45, 49), SA3) 
would create the data frame:
}
\details{
name      area          BOTTOM_DEPTH.1 BOTTOM_DEPTH.2 START_LATITUDE.1 START_LATITUDE.2
  shallow   2547.066      183            300            45.0             49.0          
  deep      2809.660      300            500            45.0             49.0
}
\seealso{
\code{\link{StrataAreas.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
