% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.r
\name{draw_theme}
\alias{draw_theme}
\title{A {ggplot2} theme for drawing maps}
\usage{
draw_theme(size = 0.25)
}
\arguments{
\item{size}{A real number specifying the line width, or size, for the panel
border around the figure. The default is \code{0.25}.}
}
\description{
A {ggplot2} theme for drawing maps
}
