% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.r
\name{draw_USEEZ}
\alias{draw_USEEZ}
\title{Bounds the figure with the limits of the U.S. EEZ}
\usage{
draw_USEEZ(lon = c(-129.15, -116.5), lat = c(31.9, 49.5))
}
\arguments{
\item{lon, lat}{Vectors of two real numbers specifying the longitudinal and
latitudinal limits for the x and y axes.}
}
\description{
Bounds the figure with the limits of the U.S. EEZ
}
