% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_growth.R
\name{est_growth}
\alias{est_growth}
\title{Estimate length-at-age using the von Bertanlaffy
parametization}
\usage{
est_growth(
  dir = NULL,
  dat,
  return_df = TRUE,
  Par = data.frame(K = 0.13, Linf = 55, L0 = 15, CV0 = 0.1, CV1 = 0.1),
  bySex = TRUE,
  estVB = TRUE,
  bins = NULL,
  sdFactor = 1,
  dopng = lifecycle::deprecated()
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{Data frame of biological data from \code{\link[=pull_bio]{pull_bio()}}}

\item{return_df}{TRUE/FALSE If set to TRUE the dat data frame is
returned with added columns for the estimated Lhat_low, Lhat_pred, and Lhat_high.}

\item{Par}{Data frame of starting parameters for K, Linf, L0, CV0, and CV2 based on the
Stock Synthesis parameterization of von Bertanlaffy growth.}

\item{bySex}{Logical to indicate if plot by sex}

\item{estVB}{Logical. Estimate vonB growth to plot against predicted length. If F, it uses the paramters in \code{parStart}.}

\item{bins}{The bins to put ages into. If NULL then simply uses the ages as recorded.}

\item{sdFactor}{The number of standard deviations to include in the
low and high calculations. The default is 1.0.}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}
}
\description{
Estimate length-at-age using the von Bertanlaffy
parametization
}
\seealso{
\code{\link[=fit_vbgrowth]{fit_vbgrowth()}}
}
\author{
Chantel Wetzel
}
