% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.AKSlope.fn.R
\name{Format.AKSlope.fn}
\alias{Format.AKSlope.fn}
\title{Deprecated function replaced by pull_catch() and pull_bio()}
\usage{
Format.AKSlope.fn(
  dir = NULL,
  datTows,
  datL = NA,
  start.year = 1997,
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{datTows}{A data frame of catch data for the
AKFSC slope survey with incorrect column names.
prior to the creation of the data warehouse.}

\item{datL}{A list of biological data (lengths and ages)
for the AKFSC slope survey with incorrect column
names prior to the creation of the data warehouse.}

\item{start.year}{The first year of data to retain within
the data frame. The first year typically used from this
survey is 1997.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Deprecated function replaced by pull_catch() and pull_bio()
}
\details{
This function is no longer used.  Please use pull_catch() and pull_bio() to get properly formatted and filtered data.
}
\author{
Chantel Wetzel
}
