% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expanded_comps.R
\name{get_expanded_comps}
\alias{get_expanded_comps}
\title{Expands the lengths up to the total stratum area then sums over strata for each year}
\usage{
get_expanded_comps(
  bio_data,
  catch_data,
  comp_bins,
  strata,
  dir = NULL,
  comp_column_name = "length_cm",
  output = c(, "full_expansion_ss3_format", "tow_expansion_only",
    "full_expansion_unformatted")[1],
  two_sex_comps = TRUE,
  input_n_method = c("stewart_hamel", "tows", "total_samples")[1],
  partition = 0,
  fleet = "Enter Fleet",
  age_low = -1,
  age_high = -1,
  age_error = "Enter Numeric",
  month = "Enter Month",
  printfolder = "forSS3",
  verbose = TRUE
)
}
\arguments{
\item{bio_data}{A data frame of length-composition data returned from
\code{\link[=pull_bio]{pull_bio()}}.}

\item{catch_data}{A data frame of catch data returned from \code{\link[=pull_catch]{pull_catch()}}.}

\item{comp_bins}{Vector of length bins to create length compositions across. Values above or below the
minimum or maximum values, respectively, are grouped into the first size or plus group size.}

\item{strata}{A data frame that defines the strata and provides the
calculated areas for each strata returned from \code{\link[=createStrataDF.fn]{createStrataDF.fn()}}.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{comp_column_name}{The column name to create composition data for. This column can be
is used to determine whether to format the composition data for length or age
compositions by looking for either age (e.g., \code{age_years}, \code{Age}, \code{best_age}) or length
(e.g., \code{Length}, \code{length}, \code{Length_cm}) in the comp_column_name. The default
is \code{Length_cm}.}

\item{output}{Switch to specify how to return the composition data where the options
are c("tow_expansion_only", "full_expansion_unformatted", "full_expansion_ss3_format").
The default is c("tow_expansion_only", "full_expansion_unformatted", "full_expansion_ss3_format")\link{3}.}

\item{two_sex_comps}{Default TRUE. If TRUE composition data will be formatted for a
Stock Synthesis two-sex model and if FALSE composition data will be formatted for a
single-sex model.}

\item{input_n_method}{Determines the default input sample size to add to
the composition data for SS3. There are three options: c("stewart_hamel", "tows",
"total_samples") where the default is "stewart_hamel".}

\item{partition}{Partition to assign the composition data based on the expected
format for Stock Synthesis. Partition of 0 indicates that the composition data
include all composition data, 1 for discarded composition data, and 2 for retained
fish only. Default of 0.}

\item{fleet}{A fleet number to assign the composition data to based on the expected
format for Stock Synthesis. Default "Enter Fleet".}

\item{age_low}{Lower age bin for all age composition data based on the expected
format for Stock Synthesis. Default value of -1 which translates to the lowest age
bin.}

\item{age_high}{Upper age bin for all age composition data based on the expected
format for Stock Synthesis. Default value of -1 which translates to the highest}

\item{age_error}{Number of ageing error vector to apply to the age data based on
Stock Synthesis. Default "Enter Age Error Vector".}

\item{month}{Month the samples were collected based on the expected format for
Stock Synthesis to determine the length/age estimate to compare to. Default
"Enter Month".}

\item{printfolder}{A string that will be appended to \code{dir}, creating a folder
where the output will be saved. If specified as \code{""},
the output will just be saved directly in \code{dir}. The default is \code{"forSS3"}.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Create expanded length composition data based on the pre-specified strata.
This function is designed to be used with catch pulled using \code{\link[=pull_catch]{pull_catch()}} and
biological data pulled using \code{\link[=pull_bio]{pull_bio()}}. The default output is formatted based
on the formatting required by Stock Synthesis.
}
\seealso{
See
\code{\link{get_input_n}} for information on input sample size calculations.
}
\author{
Chantel Wetzel and Allan Hicks
}
