% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Get json content from a URL}
\usage{
get_json(url)
}
\arguments{
\item{url}{A string containing a valid URL to pull the data from the data
warehouse.}
}
\value{
A data frame.
}
\description{
Get json content from a URL
}
\details{
Get information stored on the web in .json format using a URL. The content
is first pulled from the web as text with UTF-8 encoding. Then the text
is passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}. This workflow ensures that the URL
is not mistaken for a file name rather than web content.
}
\seealso{
See all the \verb{pull_*} functions for examples where this function is
used, e.g., \code{\link[=pull_catch]{pull_catch()}}.
}
\author{
Kelli F. Johnson
}
