% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spp_list.R
\name{get_spp_list}
\alias{get_spp_list}
\title{Get Vector of Default Species Names for US West Coast}
\usage{
get_spp_list()
}
\value{
A vector of strings specifying common names with words separated
using underscores.
}
\description{
Get Vector of Default Species Names for US West Coast
}
\details{
A standard list of common names for species of interest to the
US West Coast. This list does not reflect ecosystem or financial
importance, and instead it is strictly a list of species for which
analyses take place.
}
\examples{
get_spp_list()
}
