% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_names_long.R
\name{get_survey_names_long}
\alias{get_survey_names_long}
\title{Survey name matching function used when pull from the data warehouse}
\usage{
get_survey_names_long()
}
\description{
Survey name matching function used when pull from the data warehouse
}
\author{
Chantel Wetzel based on code by John Wallace
}
