% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAges.fn.R
\name{GetAges.fn}
\alias{GetAges.fn}
\title{Wrapper function that will read in age comp file and create age-at-length comps for SS 
This function is used for the NWFSC shelf-slope and NWFSC slope survey}
\usage{
GetAges.fn(file, headerRow, lgthBins = 1, ageBins = 1,
  nSamps = "EnterNsamps", fleet = "EnterFleet", season = 1,
  partition = 0, ageerr = "EnterAgeErr", raw = T, sep = ",")
}
\arguments{
\item{file}{The csv file with your composition data.}

\item{headerRow}{Line number indicating the header row of your csv file. (e.g., headerRow = 7)}

\item{lgthBins}{Either the interval between length bins or the actual length bins (e.g., lgthBins = 11:47)}

\item{ageBins}{Either the interval between age bins or the actual age bins (e.g, ageBins = 1:40)}

\item{nSamps}{Vector of effective sample sizes by year for data.  Value will be written in the effN column.}

\item{fleet}{Fleet number}

\item{season}{Season number}

\item{partition}{partition as defined by Stock Synthesis}

\item{ageerr}{Age error value for Stock Synthesis}

\item{raw}{raw=T/F, input to define whether or not to expand numbers in the csv file (column header "NumF" and "NumM")}

\item{sep}{Column separator in 'file'}
}
\description{
Wrapper function that will read in age comp file and create age-at-length comps for SS 
This function is used for the NWFSC shelf-slope and NWFSC slope survey
}
\author{
Allan Hicks
}
