% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAgesExcel.fn.R
\name{GetAgesExcel.fn}
\alias{GetAgesExcel.fn}
\title{Wrapper function that will read in age comp file and create age-at-length comps for SS from an excel file
This function is used for the NWFSC shelf-slope and NWFSC slope survey
Have not tested this function for reading from excel file}
\usage{
GetAgesExcel.fn(file, sheet = "AgeComps", headerRow, lgthBins = 1,
  ageBins = 1, nSamps = "EnterNsamps", fleet = "EnterFleet", season = 1,
  partition = 0, ageerr = "EnterAgeErr", raw = T)
}
\arguments{
\item{file}{The excel file with your composition data.}

\item{sheet}{Name of sheet to be read in the excel file}

\item{headerRow}{Line number indicating the header row of your csv file. (e.g., headerRow = 7)}

\item{lgthBins}{Vector of length bins (e.g., lgthBins = 11:47)}

\item{ageBins}{Vector of age bins (e.g, ageBins = 1:40)}

\item{nSamps}{Vector of effective sample sizes by year for data.  Value will be written in the effN column.}

\item{fleet}{Fleet number}

\item{season}{Season number}

\item{partition}{Partition value for composition data}

\item{ageerr}{Age error value for SS}

\item{raw}{raw=T/F, input to define whether or not to expand numbers in the csv file (column header "NumF" and "NumM")}
}
\description{
Wrapper function that will read in age comp file and create age-at-length comps for SS from an excel file
This function is used for the NWFSC shelf-slope and NWFSC slope survey
Have not tested this function for reading from excel file
}
\seealso{
\code{\link{readInExcelAgeComps.fn}}, \code{\link{SS3AgeAtLen.fn}}
}
\author{
Allan Hicks
}
