% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLFs.fn.R
\name{GetLFs.fn}
\alias{GetLFs.fn}
\title{Get length frequencies}
\usage{
GetLFs.fn(file, headerRow, lgthBins = 1, gender = 3,
  nSamps = "EnterNsamps", fleet = "EnterFleet", season = 1,
  partition = 0, NAs2zero = TRUE, sep = ",", sexRatioUnsexed = NA,
  maxSizeUnsexed = NA)
}
\arguments{
\item{file}{a CSV file containing length comp information}

\item{headerRow}{the row number of the column where the data start}

\item{lgthBins}{either the interval between length bins or the actual length bins}

\item{gender}{gender code as used by Stock Synthesis}

\item{nSamps}{value or placeholder for sample size column}

\item{fleet}{value or placeholder for fleet number column}

\item{season}{season =1}

\item{partition}{partition as defined by Stock Synthesis}

\item{NAs2zero}{determines if NA values will be changed to 0.0}

\item{sep}{column separator in 'file'}

\item{sexRatioUnsexed}{proportion used to assign unsexed fish to females
\itemize{
\item if a single number, it is only used for sizes at or below
maxSizeUnsexed, and sex ratio for size bins above that are calculated
using the number of males and females observed in the same size class
(or one lower if none available in the same bin)

\item if a vector (maybe to be implemented), it must be the same length
as the lgthBins and indicates the proportion of unsexed assigned to females
for each length bin

\item if it is NA, unsexed fish are omitted WHEN GENDER=3 (THIS IS THE DEFAULT)
}}

\item{maxSizeUnsexed}{determines the maximum size at which the sexRatioUnsexed
is applied. If sexRatioUnsexed is a vector, this is ignored.}
}
\description{
Wrapper function that calls on readInLengthComps.fn to get length frequencies
from NWFSC trawl survey data packages and then passes them to the SS3LF.fn for
formatting as required by the Stock Synthesis data file
}
\author{
Allan C. Hicks, Felipe Hurtado-Ferro, Ian G. Taylor
}
\seealso{
\code{\link{readInLengthComps.fn}}, \code{\link{SS3LF.fn}}
}

