% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN.R
\name{getN}
\alias{getN}
\title{Calculate effN input sample sizes
based on Stewart & Hamel 2014}
\usage{
getN(dir, dat, type, survey, headerRow = "default", species = "flatfish",
  printfolder = "forSS")
}
\arguments{
\item{dir}{directory}

\item{dat}{excel file name for the NWFSC or data object for the Triennial or AKSFC survey data}

\item{type}{specify whether doing "length" or "age". Used to read associatied excel sheets}

\item{survey}{survey name, options = c("NWFSC", "Tri.Shelf", "AFSC.Slope")}

\item{headerRow}{line of header row in excel file}

\item{species}{species specific value to determine the number of unique samples per tow (flatfish, shelfrock, sloperock, thorny, others, all)}

\item{printfolder}{name of the folder to create and save files. Location will be paste0(dir, printfolder)}

\item{sheet}{sheet name in excel file}
}
\description{
Calculate effN input sample sizes
based on Stewart & Hamel 2014
}
\author{
Chantel Wetzel
}
