% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpp.fn.R
\name{GetSpp.fn}
\alias{GetSpp.fn}
\title{Get Species Information}
\usage{
GetSpp.fn(species)
}
\arguments{
\item{species}{A vector of species names. The words can be separated
using spaces or underscores. Full names are not required but they will
increase the consistency of the results should partial matches return
multiple matches.}
}
\value{
A data frame with the scientific name, common name, and strata
group for each species.
}
\description{
Get the scientific name, common name, and strata group for
a vector of species.
}
\examples{
\dontrun{
GetSpp.fn(c("sablefish", "petrale"))
}
}
\author{
Kelli Faye Johnson
}
