% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSppDefault.fn.R
\name{GetSppDefault.fn}
\alias{GetSppDefault.fn}
\title{Get Vector of Default Species Names for US West Coast}
\usage{
GetSppDefault.fn()
}
\value{
A vector of strings specifying common names with words separated
using underscores.
}
\description{
A standard list of common names for species of interest to the
US West Coast. This list does not reflect ecosystem or financial
importance, and instead it is strictly a list of species for which
analyses take place.
}
\examples{
GetSppDefault.fn()
}
