% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSurveyAbb.fn.R
\name{GetSurveyAbb.fn}
\alias{GetSurveyAbb.fn}
\title{Get Survey Abbreviation}
\usage{
GetSurveyAbb.fn(surveys = "Combo")
}
\arguments{
\item{surveys}{A vector of strings specifying the survey names that you
want to get abbreviations for.}
}
\value{
A vector of strings containing a single abbreviation for each
input value in \code{surveys}. Only the first match for each survey is
returned. Values of NA were not found in the look-up matrix.
}
\description{
Get abbreviations for a vector of survey names. Standard abbreviations
do not always match the name used to extract the information from the
data warehouse, so this function will look up standarized names that
were agreed upon within the Population Ecology Program.
Partial matches are allowed, though may return multiple
values and only the first match is returned.
}
\examples{
GetSurveyAbb.fn(c("Triennial", "Combo", "shit"))
}
\seealso{
See \code{\link{createMatrix}} for a list of available surveys.
}
\author{
Kelli Faye Johnson
}
