% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTotalBiomass.fn.R
\name{GetTotalBiomass.fn}
\alias{GetTotalBiomass.fn}
\title{Wrapper function  to create the designed based indices}
\usage{
GetTotalBiomass.fn(file, headerRow, fleet = "EnterFleet", season = 1,
  outputMedian = T)
}
\arguments{
\item{file}{the BiomassAbundance csv file}

\item{headerRow}{header row in the csv file}

\item{fleet}{fleet number}

\item{season}{season number}

\item{outputMedian}{TRUE/FALSE}
}
\description{
Wrapper function  to create the designed based indices
}
\seealso{
\code{\link{readInBiomass.fn}}, \code{\link{SS3Biomass.fn}}
}
\author{
Allan Hicks
}
