% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTotalBiomassExcel.fn.R
\name{GetTotalBiomassExcel.fn}
\alias{GetTotalBiomassExcel.fn}
\title{Wrapper function  to create the designed based indices}
\usage{
GetTotalBiomassExcel.fn(file, sheet = "BiomassAbundance", headerRow,
  fleet = "EnterFleet", season = 1, outputMedian = T)
}
\arguments{
\item{file}{the excel file}

\item{sheet}{sheet name}

\item{headerRow}{header row in the csv file}

\item{fleet}{fleet number}

\item{season}{season number}

\item{outputMedian}{TRUE/FALSE}
}
\description{
Wrapper function  to create the designed based indices
}
\author{
Allan Hicks
}
