% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_tow_expansion.R
\name{label_tow_expansion}
\alias{label_tow_expansion}
\title{Label tow expanded compositions}
\usage{
label_tow_expansion(x)
}
\arguments{
\item{x}{Dataframe of tow expanded composition data created by the
\code{\link[=get_expanded_comps]{get_expanded_comps()}}.}
}
\value{
A labeled data frame.
}
\description{
Label tow expanded compositions
}
\author{
Chantel Wetzel
}
