% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_length_sampling.R
\name{plot_age_length_sampling}
\alias{plot_age_length_sampling}
\title{Plot the representativeness of age sampling based on lengths}
\usage{
plot_age_length_sampling(
  data,
  dir = NULL,
  height = 7,
  width = 7,
  xlim = lifecycle::deprecated(),
  ylim = lifecycle::deprecated()
)
}
\arguments{
\item{data}{data frame}

\item{dir}{Defaults to NULL (plot made, but not saved to file). Can alternatively be a string filename
by year and sex}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}

\item{xlim}{Deprecated with {nwfscSurvey} version 2.8.0.1 because the limits
are now automatically determined. x limits for plot, defaults to c(0,120)}

\item{ylim}{Deprecated with {nwfscSurvey} version 2.8.0.1 because the limits
are now automatically determined. y limits for plot, defaults to (0, 0.049)}
}
\description{
Plot the representativeness of age sampling based on lengths
}
\author{
Chantel Wetzel
}
