% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comps.R
\name{plot_comps}
\alias{plot_comps}
\title{This function plots frequency data as bubble plots}
\usage{
plot_comps(
  data,
  dir = NULL,
  add_save_name = NULL,
  add_0_ylim = TRUE,
  width = 10,
  height = 7
)
}
\arguments{
\item{data}{Data file object created by SS3LF.fn or SS3AF.fn}

\item{dir}{Directory to save files to}

\item{add_save_name}{Option to add text to a saved figure name. This option can
be useful if creating plots across multiple species and saving them into a single
folder. Default NULL.}

\item{add_0_ylim}{TRUE/FALSE, option to specificy if the y-axis should start at 0
or the minimum bin in the composition data file. Default TRUE.}

\item{width}{Numeric figure width in inches, defaults to 10}

\item{height}{Numeric figure height in inches, defaults to 7}
}
\description{
This function plots frequency data as bubble plots
}
\examples{
\dontrun{ plot_comps(data = LFs)}

}
\author{
Chantel Wetzel
}
