% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{This function plots cpue and length by latitude and depth}
\usage{
plot_cpue(data, dir = NULL, plot = 1:3, width = 7, height = 7)
}
\arguments{
\item{data}{Data catch file pulled using \code{\link[=pull_catch]{pull_catch()}}}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{plot}{A vector of integers to specify which plots to return. The
default is to print or save all figures, i.e., \code{plot = 1:3}. Integers
correspond to the following figures:
\enumerate{
\item log(CPUE) by depth & log(CPUE) by latitude
\item log(CPUE) by latitude and year
\item log(CPUE) by depth and year
}}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}
}
\description{
This function plots cpue and length by latitude and depth
}
\seealso{
Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_var_length_at_age}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Chantel Wetzel
}
\concept{plot_}
