% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plots the design based biomass estimates with confidence intervals}
\usage{
plot_index(
  data,
  dir = NULL,
  add_save_name = NULL,
  plot = 1:2,
  height = 7,
  width = 7,
  dpi = 300
)
}
\arguments{
\item{data}{List of design based biomass estimates created by the \code{\link[=get_design_based]{get_design_based()}}}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{add_save_name}{Option to add text to a saved figure name. This option
can be useful if creating plots across multiple species and saving them
into a single folder. The default is \code{NULL}. Note that the biomass estimate,
i.e., annual or strata, are already included in the saved name
so no need to add those here.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:2}. Integers
correspond to the following figures:
\enumerate{
\item Design based index by year.
\item Design based index by year and strata.
}}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}

\item{dpi}{The resolution to apply when saving figures.  Lower resolution values
can reduce file size which can be helpful when creating large documents
with many figures. The default is 300.}
}
\description{
Plots the design based biomass estimates with confidence intervals
}
\details{
Plots both the design based biomass estimates by year with confidence intervals
and the design based biomass estiamtes by year and strata with no confidence
intervals.
}
\author{
Chantel Wetzel
}
