% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varlenage.R
\name{plot_var_length_at_age}
\alias{plot_var_length_at_age}
\title{Plot variability of length at age}
\usage{
plot_var_length_at_age(
  data,
  age_bins,
  dir = NULL,
  main = NULL,
  two_sex = TRUE,
  height = 7,
  width = 7
)
}
\arguments{
\item{data}{A data frame of length-composition data returned from
\code{\link[=pull_bio]{pull_bio()}}.}

\item{age_bins}{Vector of integers to bin age data.Values above or below the minimum or maximum
values in the vector are grouped into the first size or plus group size, respectively.}

\item{dir}{Defaults to NULL (plot made, but not saved to file). Can alternatively be a string filename
by year and sex.}

\item{main}{Name that will be used to name the saved png}

\item{two_sex}{Logical to indicate if plot by sex. Default is TRUE and will only plot females and males.}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}
}
\description{
Plot variability of length at age
}
\seealso{
Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Chantel Wetzel
}
\concept{plot_}
