% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varlenage.R
\name{plot_varlenage}
\alias{plot_varlenage}
\title{Plot variability of length at age}
\usage{
plot_varlenage(
  dir = NULL,
  dat,
  main = NULL,
  Par = data.frame(K = 0.13, Linf = 55, L0 = 15, CV0 = 0.1, CV1 = 0.1),
  bySex = TRUE,
  estVB = TRUE,
  bins = NULL,
  legX = "bottomleft",
  legY = NULL,
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{The data loaded from \code{\link[=pull_bio]{pull_bio()}}}

\item{main}{Name that will be used to name the saved png}

\item{Par}{Starting parameters for K, Linf, L0, CV0, and CV2 based on the
Stock Synthesis parameterization of von Bertanlaffy growth.}

\item{bySex}{Logical to indicate if plot by sex}

\item{estVB}{Logical. Estimate vonB growth to plot against predicted length.
If F, it uses the paramters in \code{parStart}.}

\item{bins}{The bins to put ages into. If NULL then simply uses the ages as
recorded.}

\item{legX}{legend location for x axis, defaults to "bottomleft"}

\item{legY}{legend location for y axis, defaults to NULL}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots.}
}
\description{
Plot variability of length at age
}
\details{
Plots the standard deviation and coefficient of variation of age at observed
and predicted length
}
\author{
Chantel Wetzel
}
