% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_westcoast.R
\name{plot_westcoast}
\alias{plot_westcoast}
\title{Plot the United States west coast using {ggplot2}}
\usage{
plot_westcoast()
}
\value{
A {ggplot2} object.
}
\description{
Plot the United States west coast such that it can used as a standalone
figure or as a base to plot data points on.
}
\examples{
\dontrun{
if (require("maps") && require("mapproj")) {
  map_object <- plot_westcoast()
  print(map_object)
  ggplot2::ggsave(
    filename = "myfilename.png",
    plot = map_object
  )
}
}
}
\seealso{
\itemize{
\item \code{plot_cpue_map()} uses this function as the base map.
}

Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_varlenage}()},
\code{\link{plot_weight_length}()}
}
\author{
Kelli F. Johnson
}
\concept{plot_}
