% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBio.fn.R
\name{PlotBio.fn}
\alias{PlotBio.fn}
\title{Plots the biomass with confidence intervals}
\usage{
PlotBio.fn(dir = NULL, dat, CI = 0.95, scalar = 1e+06, gap = 0.03,
  ylab = "Biomass ('000 mt)", xlab = "Year", main = NULL, ylim = NULL,
  add = FALSE, col = "black", dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory to save the file}

\item{dat}{object created by the GetTotalBiomass.fn}

\item{CI}{confidence interval}

\item{scalar}{simply the divisor for the biomass}

\item{gap}{a value that introduces a slight gap between the point estimate and the start of the line for the CI. A gap too large will invert the CI, making it look huge. You should know when this happens}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{main}{plot label}

\item{ylim}{y-limits}

\item{add}{add additional line to plot}

\item{col}{color}

\item{dopng}{save the plot as a png inside plots folder}

\item{...}{Additional arguments for the plots}
}
\description{
Plots the biomass with confidence intervals
}
\author{
Allan Hicks and John Wallace
}
