% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioStrata.fn.R
\name{PlotBioStrata.fn}
\alias{PlotBioStrata.fn}
\title{Plots the design-based biomass estimates by strata with confidence intervals}
\usage{
PlotBioStrata.fn(
  dir = NULL,
  dat,
  CI = 0.95,
  scalar = 1000000,
  gap = 0.03,
  ylab = "Biomass ('000 mt)",
  xlab = "Year",
  survey.name = NULL,
  strata.names = NULL,
  ylim = NULL,
  sameylim = FALSE,
  add = FALSE,
  mfrow.in = NULL,
  col = "black",
  pch.col = "black",
  pch.type = 16,
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{Data frame created by the \code{\link[=Biomass.fn]{Biomass.fn()}}}

\item{CI}{A numerical value that specifies the confidence interval to return.
Values should be between 0.01 to 0.99.}

\item{scalar}{simply the divisor for the biomass}

\item{gap}{a value that introduces a slight gap between the point estimate and the start of the line for the CI. A gap too large will invert the CI, making it look huge. You should know when this happens}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{survey.name}{if specified will name the output png file using this name}

\item{strata.names}{custom strata names, if not specified will use the defined names from \code{\link[=CreateStrataDF.fn]{CreateStrataDF.fn()}}}

\item{ylim}{y-limits}

\item{sameylim}{Boolean, whether to include same y-limits or not. Defaults to FALSE}

\item{add}{add additional line to plot}

\item{mfrow.in}{option to specify the mfrow for plotting}

\item{col}{color}

\item{pch.col}{Color as string, defaults to "black"}

\item{pch.type}{Numeric pch type, defaults to 16}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots}
}
\description{
Plots the design-based biomass estimates by strata with confidence intervals
}
\author{
Chantel Wetzel, Allan Hicks, and John Wallace
}
