% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreqData.fn.R
\name{PlotFreqData.fn}
\alias{PlotFreqData.fn}
\title{Plot length or age compositions by year in bubble plots}
\usage{
PlotFreqData.fn(
  dir = NULL,
  dat,
  inch = 0.15,
  ylab = "Bins",
  xlab = "Year",
  zero2NAs = TRUE,
  main = NULL,
  xlim = NULL,
  ymax = NULL,
  dopng = lifecycle::deprecated(),
  w = 7,
  h = 7,
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{Object created by \code{\link[=SurveyLFs.fn]{SurveyLFs.fn()}} or \code{\link[=SurveyAFs.fn]{SurveyAFs.fn()}}}

\item{inch}{input to the symbols plot: TRUE, FALSE or a positive number.}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{zero2NAs}{A logical specifying if \code{NA}s should be changed to zeros.
The default is \code{TRUE}.}

\item{main}{main plot text}

\item{xlim}{x-limit values}

\item{ymax}{Value used to truncate y-axis, defaults to NULL}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{w}{Numeric figure width, defaults to 7}

\item{h}{Numeric figure height, defaults to 7}

\item{...}{Additional arguments for the plots}
}
\description{
Plot length or age compositions by year in bubble plots
}
\author{
Allan Hicks and Chantel Wetzel
}
