% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreqDataStrata.fn.R
\name{plotFreqDataStrata.fn}
\alias{plotFreqDataStrata.fn}
\title{I have not tested this function to make sure it works correctly: Chantel Wetzel
This function plots frequency data as bubble plots when strata instead of years are used
You may want to change all zeros to NA's so that those observations are not plotted.
If you don't then set zero2NAs=F}
\usage{
plotFreqDataStrata.fn(dat, inch = 0.15, ylab = "Bins", xlab = "",
  zero2NAs = T, ...)
}
\arguments{
\item{dat}{object created by SS3LF.fn or SS3AF.fn}

\item{inch}{input to the symbols plot: TRUE, FALSE or a positive number.}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{zero2NAs}{T/F change 0 values to NA}
}
\description{
I have not tested this function to make sure it works correctly: Chantel Wetzel
This function plots frequency data as bubble plots when strata instead of years are used
You may want to change all zeros to NA's so that those observations are not plotted.
If you don't then set zero2NAs=F
}
\author{
Allan Hicks
}
