% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLFs.fn.R
\name{plotLFs.fn}
\alias{plotLFs.fn}
\title{A wrapper for the plotFreqData 
because I originally called it plotLFs.fn and want to keep it compatible with already written analyses}
\usage{
plotLFs.fn(dat, inch = 0.15, ylab = "Bins", xlab = "Year", zero2NAs = T,
  ...)
}
\arguments{
\item{dat}{object created by SS3LF.fn or SS3AF.fn}

\item{inch}{input to the symbols plot: TRUE, FALSE or a positive number.}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{zero2NAs}{T/F change 0 values to NA}
}
\description{
A wrapper for the plotFreqData 
because I originally called it plotLFs.fn and want to keep it compatible with already written analyses
}
\author{
Allan Hicks
}
