% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMap.fn.R
\name{PlotMap.fn}
\alias{PlotMap.fn}
\title{Create maps with catch-per-unit-effort}
\usage{
PlotMap.fn(
  dir = NULL,
  dat,
  main = NULL,
  dopng = lifecycle::deprecated(),
  plot = 1:2
)
}
\arguments{
\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{An object created by \code{\link[=PullCatch.fn]{PullCatch.fn()}}.}

\item{main}{A string that will be prepended to the name of the saved png
(i.e., "NWFSC" results in a file called "NWFSC_CPUE_Map.png").}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{plot}{A vector of integers specifying the selection of a subset of
plots: 1 = coastwide across all years, 2 = coastwide by year.}
}
\description{
Create the following figures:
\enumerate{
\item CPUE plot of the data across all years and/or
\item CPUE plot of the data by year.
}
}
\author{
Chantel R. Wetzel
}
