% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexRatio.fn.R
\name{plotSexRatio.fn}
\alias{plotSexRatio.fn}
\title{Function to plot sex ratio}
\usage{
plotSexRatio.fn(len, fn = median, circleSize = 0.1, ...)
}
\arguments{
\item{len}{data object}

\item{fn}{value to calculate from the data (e.g., median, mean)}

\item{circleSize}{circle size}
}
\description{
Function to plot sex ratio
}
\author{
Allan Hicks
}
