% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarLengthAtAge.fn.R
\name{PlotVarLengthAtAge.fn}
\alias{PlotVarLengthAtAge.fn}
\title{Plot variability of length at age}
\usage{
PlotVarLengthAtAge.fn(
  dir = NULL,
  dat,
  main = NULL,
  ageBin = 1,
  bySex = T,
  parStart = c(52, 0.09, 1),
  estVB = T,
  bins = NULL,
  legX = "bottomleft",
  legY = NULL,
  verbose = verbose,
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{A data frame of length-composition data returned from
\code{\link[=pull_bio]{pull_bio()}}.}

\item{main}{Name that will be used to name the saved png}

\item{ageBin}{Currently fixed at 1, so a moot parameter}

\item{bySex}{Logical to indicate if plot by sex}

\item{parStart}{Vector of starting parameters for Linf, k, and t0 in VonB estimation}

\item{estVB}{Logical. Estimate vonB growth to plot against predicted length. If F, it uses the parameters in \code{parStart}.}

\item{bins}{The bins to put ages into. If NULL then simply uses the ages as recorded.}

\item{legX}{legend location for x axis, defaults to "bottomleft"}

\item{legY}{legend location for y axis, defaults to NULL}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots}
}
\description{
Plots the SD and CV of age at observed and predicted length
}
\seealso{
Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_bio_patterns}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_var_length_at_age}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Allan Hicks and Chantel Wetzel
}
\concept{plot_}
