% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_bio.R
\name{pull_bio}
\alias{pull_bio}
\title{Pull biological data from the NWFSC data warehouse
The website is: https://www.webapps.nwfsc.noaa.gov/data
This function can be used to pull a single species or all observed species
In order to pull all species leave common_name or sci_name as NULL}
\usage{
pull_bio(
  common_name = NULL,
  sci_name = NULL,
  years = c(1980, 2050),
  survey = NULL,
  dir = NULL,
  convert = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{common_name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{sci_name} argument if you know the latin name.}

\item{sci_name}{A character entry with the desired scientific name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{common_name} argument if you know the common name.}

\item{years}{An integer vector of length two with the
range of years to pull data for.}

\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.
}}

\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{convert}{TRUE/FALSE to convert column names to first letter uppercase
which aligns with the expected names in data processing functions.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Pull biological data from the NWFSC data warehouse
The website is: https://www.webapps.nwfsc.noaa.gov/data
This function can be used to pull a single species or all observed species
In order to pull all species leave common_name or sci_name as NULL
}
\examples{
\dontrun{
# SurveyName is only arg that has to be specified
bio_dat <- PullBio.fn(SurveyName = "NWFSC.Combo")

# Example with specified common name
bio_dat <- PullBio.fn(Name = "vermilion rockfish",
SurveyName = "NWFSC.Combo")

# Example with specified scientific name
bio_dat <- PullBio.fn(SciName = "Eopsetta jordani",
SurveyName = "NWFSC.Combo")

# Example with multiple names
bio_dat <- PullBio.fn(SciName = c("Sebastes aurora","Eopsetta jordani"),
SurveyName = "NWFSC.Combo")
}

}
\author{
Chantel Wetzel
}
