% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hkl_cache.R
\name{pull_hkl_cache}
\alias{pull_hkl_cache}
\title{Pull NWFSC Hook-and-Line Survey data from pinned data list on posit connect}
\usage{
pull_hkl_cache(
  years = c(1970, 2050),
  common_name = NULL,
  dir = NULL,
  return_positive_sites_only = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{years}{An integer vector of length two with the
range of years to pull data for (e.g., c(2003, 2024)).
Vector can not contain -Inf or Inf.}

\item{common_name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{sci_name} argument if you know the latin name.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}

\item{return_positive_site_only}{Logical that specifies whether data from sites
that have observed common_name at least once across the time series are
returned.The default is \code{TRUE}.}
}
\value{
Returns a list of hook-and-line catch and biological data
}
\description{
This function can be used to return pinned data for NWFSC Hook-and-Line Survey.
If a \code{common_name} is provided the data will be formatted and filtered for
that species.  A list is returned with catch and biological data.
}
\seealso{
Other data pulling functions:
\code{\link[=pull_bio]{pull_bio()}},
\code{\link[=pull_bio_cache]{pull_bio_cache()}},
\code{\link[=pull_biological_samples]{pull_biological_samples()}},
\code{\link[=pull_catch]{pull_catch()}},
\code{\link[=pull_catch_cache]{pull_catch_cache()}},
\code{\link[=pull_gemm]{pull_gemm()}},
\code{\link[=pull_haul]{pull_haul()}},
\code{\link[=pull_haul_cache]{pull_haul_cache()}}
}
\author{
Chantel Wetzel
}
\concept{data pulling functions}
