% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/pull_spp.R
\docType{data}
\name{pull_spp}
\alias{pull_spp}
\title{Sql pull of species names}
\format{
A data frame with four columns,
latin, common, common_name, and scientific_name.
}
\usage{
pull_spp

pull_spp()
}
\description{
A static version of the sql call to pull species taxonomic information
from the data warehouse.

Pull common name and scientific name information from the
data warehouse.
}
\examples{
\dontrun{
spp <- pull_spp()
}

}
\seealso{
\code{\link{pull_spp}} for more information.
}
\author{
Curt Whitmire

Kelli Faye Johnson
}
\keyword{datasets}
