% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.fn.R
\name{PullCatch.fn}
\alias{PullCatch.fn}
\title{Pull catch data from the NWFSC data warehouse}
\usage{
PullCatch.fn(
  Name = NULL,
  SciName = NULL,
  YearRange = c(1980, 5000),
  SurveyName = NULL,
  SaveFile = lifecycle::deprecated(),
  Dir = NULL,
  verbose = TRUE,
  sample_types = c("NA", NA, "Life Stage", "Size")[1:2]
)
}
\arguments{
\item{Name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{SciName} argument if you know the latin name.}

\item{SciName}{A character entry with the desired scientific name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{Name} argument if you know the common name.}

\item{YearRange}{An integer vector of length two with the
range of years to pull data for.}

\item{SurveyName}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (not yet working),
\item NWFSC.Hook.Line (not yet working),
\item NWFSC.Video,
\item Triennial.Canada
}

Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.}

\item{SaveFile}{Deprecated with {nwfscSurvey} 2.3. Output will be save automatically
if the Dir input is specified.}

\item{Dir}{The directory where you want the output file to be saved.
The name of the file within \code{Dir} will start with Catch_ and end with .rdata.
Default NULL which will not save an output file.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}

\item{sample_types}{A character vector of sample types, i.e.,
\code{"statistical_partition_dim"}, that you would like to keep. The default is
to only keep \code{NA} values, both real and character NA. But, for some
instances you may want to keep Life Stage and Size samples. The majority
of samples with \code{"statistical_partition_dim"} of Size and Life Stage are
Pacific hake and should not be considered different than regular survey
samples. The other types of samples that may be designated Life Stage are
egg cases that can be caught and identified for select elasmobranch
species. These type of samples would not be included in the data to estimate
indices of abundance and are omitted by default.}
}
\description{
Pull catch data from the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
for a single species or all observed species, where the latter is specified
by leaving both \code{Name = NULL} and \code{SciName = NULL}.
}
\details{
The data available in the warehouse are cleaned pior to being downloaded
with the intent that they provide the best available information for use
in an index-standardization procedure. The removed samples may be of use
to others with a less-restrictive goal than producing an index of abundance.
For example, life-stage samples are excluded because they are not collected
using the same protocols as standard samples.
To download all data, we currently recommend going to the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
and using the csv link to extract data for a single species at a time.
In the future, we hope to add functionality to this package such that
downloading all data can be done easily within this function.
See \href{https://github.com/pfmc-assessments/nwfscSurvey/issues/43}{Issue #43}
for more information.
}
\examples{
\dontrun{
# SurveyName is only arg that has to be specified
dat <- PullCatch.fn(SurveyName = "NWFSC.Combo")

# Example with specified common name
catch_dat <- PullCatch.fn(Name = "vermilion rockfish",
SurveyName = "NWFSC.Combo")

# Example with specified scientific name
catch_dat <- PullCatch.fn(SciName = "Eopsetta jordani",
SurveyName = "NWFSC.Combo")

# Example with multiple names
catch_dat <- PullBio.fn(Name = c("vermilion rockfish",
"vermilion and sunset rockfish"), SurveyName = "NWFSC.Combo")

}

}
\author{
Chantel Wetzel (maintainer) based on code by John Wallace
}
