% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullSpp.fn.R
\name{PullSpp.fn}
\alias{PullSpp.fn}
\title{Pull Species Names From the Warehouse}
\usage{
PullSpp.fn()
}
\description{
Pull common name and scientific name information from the
data warehouse.
The website is https://www.webapps.nwfsc.noaa.gov/data
}
\examples{
\dontrun{
spp <- PullSpp.fn()
}

}
\author{
Kelli Faye Johnson
}
