% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PullSpp}
\alias{PullSpp}
\title{Sql pull of species names}
\format{
A data frame with four columns,
latin, common, common_name, and scientific_name.
}
\usage{
PullSpp
}
\description{
A static version of the sql call to pull species taxonomic information
from the data warehouse.
}
\seealso{
\code{\link{PullSpp.fn}} for more information.
}
\author{
Curt Whitmire
}
\keyword{datasets}
