% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataFromExcel.fn.R
\name{readDataFromExcel.fn}
\alias{readDataFromExcel.fn}
\title{Reads in the data from an Excel sheet
the column names of the data are on row number headerRow
this function is meant to be used by more specific functions below
originally written in 2009 by Allan Hicks
note that in later, 64-bit versions of R, the RODBC package is not working
updated in 2018 by Chantel Wetzel to use the readxl package}
\usage{
readDataFromExcel.fn(file, sheet, headerRow)
}
\arguments{
\item{file}{excel file}

\item{sheet}{sheet name}

\item{headerRow}{line of header row in excel file}
}
\description{
Reads in the data from an Excel sheet
the column names of the data are on row number headerRow
this function is meant to be used by more specific functions below
originally written in 2009 by Allan Hicks
note that in later, 64-bit versions of R, the RODBC package is not working
updated in 2018 by Chantel Wetzel to use the readxl package
}
\author{
Allan Hicks & Chantel Wetzel
}
