% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInAgeComps.fn.R
\name{readInAgeComps.fn}
\alias{readInAgeComps.fn}
\title{Reads in the stratum numbers from the Excel file provided by Beth (AgeComps Sheet)
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet}
\usage{
readInAgeComps.fn(file, headerRow = 7, sep = ",",
  colNames = c("SpeciesCode", "Species", "Year", "Project", "AreaName",
  "MinStratumDepth", "MaxStratumDepth", "Length", "Age", "NumF", "NumM",
  "NumUnsexed", "LengthedAgeTally", "AgeTallyF", "AgeTallyM", "AgeTallyU"))
}
\arguments{
\item{file}{csv file}

\item{headerRow}{line of header row in excel file}

\item{sep}{seperator}

\item{colNames}{column names the overwrite the one in the csv file}
}
\description{
Reads in the stratum numbers from the Excel file provided by Beth (AgeComps Sheet)
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet
}
\author{
Allan Hicks
}
