% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInAges.EWC.fn.R
\name{ReadInAges.EWC.fn}
\alias{ReadInAges.EWC.fn}
\title{Reads in the West Coast Trienial survey data and filters the data into what is necessary
It reads in data and makes sure only the species necessary are kept
may want to keep NA (blank in Excel) to select the zero tows
removeCAN is a flag if you want tows in Canadian waters removed}
\usage{
ReadInAges.EWC.fn(dataFile, directory, species = c(NA), removeCAN = T,
  verbose = F)
}
\arguments{
\item{dataFile}{data file name}

\item{directory}{working directory where the dataFile is located}

\item{species}{species name to extract}

\item{removeCAN}{removes Canadian hauls based on the AFSCforeign_hauls.rda file}

\item{verbose}{write out comments}
}
\description{
Necessary column names
   SPECIES_CODE
   AGE
}
\author{
Allan Hicks
}
