% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInAges.fn.R
\name{ReadInAges.fn}
\alias{ReadInAges.fn}
\title{Deprecated function replaced by pull_bio()}
\usage{
ReadInAges.fn(dat, subset_years = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{data file name}

\item{subset_years}{specify the years to retain, default is NULL which will
provide 1977, alternative input would be 1980:2002 to remove only 1977.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Deprecated function replaced by pull_bio()
}
\author{
Allan Hicks and Chantel Wetzel
}
