% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInBiomass.EWC.fn.R
\name{ReadInBiomass.EWC.fn}
\alias{ReadInBiomass.EWC.fn}
\title{Reads in the triennial survey data and filters the data into what is necessary
It reads in catch and station data and makes sure only the species necessary are kept
may want to keep NA (blank in Excel) to select the zero tows
removeCAN is a flag if you want tows in Canadian waters removed}
\usage{
ReadInBiomass.EWC.fn(dataFile, directory, species = c(NA), removeCAN = T,
  verbose = F)
}
\arguments{
\item{dataFile}{data file name}

\item{directory}{working directory where the dataFile is located}

\item{species}{species name to extract}

\item{removeCAN}{removes Canadian hauls based on the AFSCforeign_hauls.rda file}

\item{verbose}{write out comments}
}
\description{
Necessary column names
   SPECIES_CODE
   WEIGHT in kg
   DISTANCE_FISHED is in km
   NET_WIDTH is in m
}
\author{
Allan Hicks
}
