% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInBiomass.fn.R
\name{readInBiomass.fn}
\alias{readInBiomass.fn}
\title{Reads in the stratum biomasses from the text file saved from the Excel worksheet
headerRow is the line number where the data start (use to skip over header lines)
you can keep your column names or use the simplified names I provide.
Make sure that these match what is in your Excel spreadsheet or make sense!}
\usage{
readInBiomass.fn(filename, headerRow = 1, sep = ",",
  colNames = c("Species", "ScientificName", "SpeciesCode", "Year", "Project",
  "StrataAreaVersion", "AreaSetId", "AreaName", "SouthernLatitude",
  "NorthernLatitude", "DepthStrataSet", "MinStratumDepth", "MaxStratumDepth",
  "StratumArea", "Biomass", "Abundance", "CpueWeightVar", "CpueCountVar",
  "BiomassVar", "CV", "N", "Nbio", "Npos", "NbioPos"))
}
\arguments{
\item{filename}{csv file}

\item{headerRow}{line of header row in excel file}

\item{sep}{seperator}

\item{colNames}{column names the overwrite the one in the csv file}
}
\description{
Reads in the stratum biomasses from the text file saved from the Excel worksheet
headerRow is the line number where the data start (use to skip over header lines)
you can keep your column names or use the simplified names I provide.
Make sure that these match what is in your Excel spreadsheet or make sense!
}
\author{
Allan Hicks
}
