% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInExcelLengthComps.fn.R
\name{readInExcelLengthComps.fn}
\alias{readInExcelLengthComps.fn}
\title{Reads in the stratum numbers from the Excel file provided by Beth (LengthComps Sheet)
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet.}
\usage{
readInExcelLengthComps.fn(file, sheet = "LengthComps", headerRow = 7)
}
\arguments{
\item{file}{excel file name}

\item{sheet}{sheet name in excel file}

\item{headerRow}{line of header row in excel file}
}
\description{
Reads in the stratum numbers from the Excel file provided by Beth (LengthComps Sheet)
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet.
}
\seealso{
\code{\link{readDataFromExcel.fn}}
}
\author{
Allan Hicks
}
