% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInLengthComps.fn.R
\name{readInLengthComps.fn}
\alias{readInLengthComps.fn}
\title{Reads in the stratum numbers from a csv file (usually converted from Excel file provided by Beth (LengthComps Sheet))
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet.}
\usage{
readInLengthComps.fn(file, headerRow = 7, sep = ",",
  colNames = c("SpeciesCode", "ScientificName", "Species", "Year", "Project",
  "AreaSetIdentifier", "AreaName", "DepthStrataSet", "MinStratumDepth",
  "MaxStratumDepth", "Length", "NumF", "NumM", "NumUnsexed"))
}
\arguments{
\item{file}{csv file name}

\item{headerRow}{line of header row in excel file}

\item{sep}{seperator}

\item{colNames}{column names}
}
\description{
Reads in the stratum numbers from a csv file (usually converted from Excel file provided by Beth (LengthComps Sheet))
headerRow is the row number of the column where the data start
it doesn't read in the column names correctly, so I put in simplified names. 
Make sure that these match what is in your Excel spreadsheet.
}
\author{
Allan Hicks
}
