% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInLengths.EWC.fn.R
\name{ReadInLengths.EWC.fn}
\alias{ReadInLengths.EWC.fn}
\title{Reads in the triennial survey data and filters the data into what is necessary
It reads in catch and station data and makes sure only the species necessary are kept
### may want to keep NA (blank in Excel) to select the zero tows
removeCAN is a flag if you want tows in Canadian waters removed
   need the file called foreign_hauls.csv}
\usage{
ReadInLengths.EWC.fn(dataFile, directory, species = c(NA), removeCAN = T,
  verbose = F)
}
\arguments{
\item{dataFile}{csv file name}

\item{directory}{working directory where the dataFile is located}

\item{species}{species name}

\item{removeCAN}{switch to remove Canadian hauls}

\item{verbose}{print out comments}
}
\description{
Necessary column names
   SPECIES_CODE
   LENGTH
}
\author{
Allan Hicks
}
