% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInLengths.R
\name{ReadInLengths.fn}
\alias{ReadInLengths.fn}
\title{Deprecated function replaced by pull_bio()}
\usage{
ReadInLengths.fn(dat, verbose = TRUE)
}
\arguments{
\item{dat}{data file name}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Deprecated function replaced by pull_bio()
}
\author{
Allan Hicks and Chantel Wetzel
}
