% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SA3}
\alias{SA3}
\title{A data set for determining strata areas (hectares)}
\format{
A data frame.
}
\usage{
SA3
}
\description{
Data necessary for determining the amount of area a strata covers to
expand estimates of biomass to be absolute.
The data includes the following columns:
\itemize{
\item SUBAREA_ID
\item MIN_DEPTH_M
\item MAX_DEPTH_M
\item AREA_HECTARES
\item SUBAREA_SET_ID
\item MIN_LAT_DD
\item MAX_LAT_DD
}
}
\author{
Curt Whitmire
}
\keyword{datasets}
