% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SA3_v2021.1}
\alias{SA3_v2021.1}
\title{A data set for determining strata areas (hectares)}
\format{
A data frame.
}
\usage{
SA3_v2021.1
}
\description{
Data necessary for determining the amount of area a strata covers to
expand estimates of biomass to be absolute. The \code{SA3} data set
was updated in March of 2021 to allow for splitting areas at the
40.16667, a common management line.
The data includes the following columns:
\itemize{
\item OBJECT_ID
\item MIN_DEPTH_M
\item MAX_DEPTH_M
\item MIN_LAT_DD
\item MAX_LAT_DD
\item AREA_HECTARES
\item INPFC_AREA
}
}
\author{
Curt Whitmire
}
\keyword{datasets}
