% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rdata.r
\name{save_rdata}
\alias{save_rdata}
\title{Save an object in an \code{.rdata} file}
\usage{
save_rdata(x, dir = NULL, name_base, verbose = TRUE)
}
\arguments{
\item{x}{An object.}

\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{name_base}{A string that will be appended to with the system time and
the file extension (i.e., \code{".rdata"}).}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
Nothing is returned, instead a file is saved in \code{dir}.
}
\description{
Save an object in an \code{.rdata} file
}
