% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SexRatio.fn.R
\name{SexRatio.fn}
\alias{SexRatio.fn}
\title{Sex ratio function that assigns a sex to unsexed fish when creatin length or age comps}
\usage{
SexRatio.fn(datB, sexRatioUnsexed, maxSizeUnsexed)
}
\arguments{
\item{datB}{the read in length comps by the ReadInLengths.EWC.fn function}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}
}
\description{
Sex ratio function that assigns a sex to unsexed fish when creatin length or age comps
}
\author{
Allan Hicks and Chantel Wetzel
}
