% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS3AF.fn.R
\name{SS3AF.fn}
\alias{SS3AF.fn}
\title{Wrapper function to create age frequency data in format for Stock Synthesis}
\usage{
SS3AF.fn(dir, Age.df, gender, ageBins, agelow = -1, agehigh = -1,
  ageErr = "NA", printfolder = "forSS", ...)
}
\arguments{
\item{dir}{directory to save the file}

\item{Age.df}{age data frame}

\item{gender}{gender for SS}

\item{ageBins}{defined age bins}

\item{agelow}{SS value}

\item{agehigh}{SS value}

\item{ageErr}{age error vector value}

\item{printfolder}{folder where the length comps will be saved}

\item{...}{additional inputs}
}
\description{
Wrapper function to create age frequency data in format for Stock Synthesis
}
\seealso{
\code{\link{SS3.LF.fn}}
}
\author{
Allan Hicks
}
