% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS3Biomass.fn.R
\name{SS3Biomass.fn}
\alias{SS3Biomass.fn}
\title{This function outputs a dataframe in the abundance index format needed for SS3
it calculates the standard error of log(B) by first finding the cv of B
It can output the median biomass or the mean biomass. Median biomass is bias corrected so that log(Bmedian)=mean(log(B))}
\usage{
SS3Biomass.fn(bio, fleet = "EnterFleet", season = 1, outputMedian = T)
}
\arguments{
\item{bio}{object created by the readInBiomass.fn}

\item{fleet}{fleet number}

\item{season}{season}

\item{outputMedian}{TRUE/FALSE}
}
\description{
This function outputs a dataframe in the abundance index format needed for SS3
it calculates the standard error of log(B) by first finding the cv of B
It can output the median biomass or the mean biomass. Median biomass is bias corrected so that log(Bmedian)=mean(log(B))
}
\author{
Allan Hicks & Chantel Wetzel
}
