% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS3LF.fn.R
\name{SS3LF.fn}
\alias{SS3LF.fn}
\title{Calculates proportion at length and reformats into SS# format
Gender=0: sexes combined and entered in female placeholders. Male values ignored
Gender=1: females only. Male values ignored
Gender=2: males only. Female values ignored.
Gender=3: both sexes. Proportions over males and females sum to 1. See below for how unsexed are treated
lgthBins is either the interval between length bins or the actual length bins
note that 0 and Inf are tacked on the ends to account for lengths outside the interval
The largest length bin includes all lengths greater
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS3, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
The NAs2zero determines if NA values will be changed to 0.0.
You will want this to be T when inputting into SS3, but may want it to be false when plotting
Note that an NA means there were no males and females recorded. Therefore you may get a zero if only one sex was recorded
Sex Ratio for unsexed
sexRatioUnsexed is the proportion used to assign unsexed fish to females
  -- if a single number, it is only used for sizes at or below maxSizeUnsexed, and sex ratio for size bins above that are calculated using the 
      number of males and females observed in the same size class (or one lower if none available in the same bin)
  -- if a vector (maybe to be implemented), it must be the same length as the lgthBins and indicates the proportion of unsexed assigned to females for each length bin
  -- if it is NA, unsexed fish are omitted WHEN GENDER=3 (THIS IS THE DEFAULT)
maxSizeUnsexed determines the maximum size at which the sexRatioUnsexed is applied. If sexRatioUnsexed is a vector, this is ignored}
\usage{
SS3LF.fn(dir, len, lgthBins = 1, gender = 3, nSamps = "EnterNsamps",
  fleet = "EnterFleet", season = 1, partition = 0, NAs2zero = T,
  sexRatioUnsexed = NA, maxSizeUnsexed = NA, printfolder = "forSS",
  remove999 = TRUE)
}
\arguments{
\item{dir}{directory to save the file}

\item{len}{object}

\item{lgthBins}{Either the interval between length bins or the actual length bins (e.g., lgthBins = 11:47)}

\item{gender}{gender code value for Stock Synthesis (0 = sexes combined, 1 = females only, 2 = males only, 3 = both sexes, females then males). Selecting gender 0 will give the user all fish combined males + females + unsexed.}

\item{nSamps}{effN}

\item{fleet}{Fleet number}

\item{season}{Season number}

\item{partition}{partition as defined by Stock Synthesis (0 = whole catch in weight, 1 = discards in weight, 2 = retained catch in weight). Survey should have a value of 0.}

\item{NAs2zero}{change NA value to 0 in the ouput length comps to use in SS}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{printfolder}{folder where the length comps will be saved}

\item{remove999}{the output object by the function will have the 999 column combined with the first length bin}
}
\description{
Calculates proportion at length and reformats into SS# format
Gender=0: sexes combined and entered in female placeholders. Male values ignored
Gender=1: females only. Male values ignored
Gender=2: males only. Female values ignored.
Gender=3: both sexes. Proportions over males and females sum to 1. See below for how unsexed are treated
lgthBins is either the interval between length bins or the actual length bins
note that 0 and Inf are tacked on the ends to account for lengths outside the interval
The largest length bin includes all lengths greater
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS3, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
The NAs2zero determines if NA values will be changed to 0.0.
You will want this to be T when inputting into SS3, but may want it to be false when plotting
Note that an NA means there were no males and females recorded. Therefore you may get a zero if only one sex was recorded
Sex Ratio for unsexed
sexRatioUnsexed is the proportion used to assign unsexed fish to females
  -- if a single number, it is only used for sizes at or below maxSizeUnsexed, and sex ratio for size bins above that are calculated using the 
      number of males and females observed in the same size class (or one lower if none available in the same bin)
  -- if a vector (maybe to be implemented), it must be the same length as the lgthBins and indicates the proportion of unsexed assigned to females for each length bin
  -- if it is NA, unsexed fish are omitted WHEN GENDER=3 (THIS IS THE DEFAULT)
maxSizeUnsexed determines the maximum size at which the sexRatioUnsexed is applied. If sexRatioUnsexed is a vector, this is ignored
}
\author{
Allan Hicks and Chantel Wetzel
}
