% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnot.R
\name{stopifnotcolumn}
\alias{stopifnotcolumn}
\title{Stop if column name is not present in the data}
\usage{
stopifnotcolumn(data, string)
}
\arguments{
\item{data}{A data frame with column names.}

\item{string}{A character string that you want to know if it exists in
the column names of \code{data}.}
}
\value{
\code{TRUE} is invisibly returned if the column name is found and
an informative \link{stop} call is initiated if the column is not present.
}
\description{
Similar in function to \link{stopifnot} but specific to a column name
that you wish to double check if it is present in your data or not.
}
