% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataAreas.fn.R
\name{StrataAreas.fn}
\alias{StrataAreas.fn}
\title{Determine strata area.}
\usage{
StrataAreas.fn(strat.df, df = SA3, convertFactor = 100)
}
\arguments{
\item{strat.df}{A dataframe describing the strata name, area, and
boundaries.  Boundaries are determined by latitude and bottom depth. See
Details.}

\item{df}{the dataframe to calculate areas for the West Coast}

\item{convertFactor}{Multiplier on the area in SA3. Default = 0.01 to
convert hectares to square km.}
}
\value{
Returns the dataframe input with a column called area.
}
\description{
calculates the area of your strata using John Wallace's SA3 file
a convertFactor of 0.01 convert hectares to km2
}
\details{
strat.df is a dataframe with a column called "name" and two columns for
each stratum variable (START_LATITUDE and BOTTOM_DEPTH) indicating the low
and high bounds.  The columns for each stratum variable should be named with
the stratum variable name followed by .1 for the low bound and .2 for the
high bound. For example, this is what a strat.df would look like (as called
in the example below).
\preformatted{
  name  area    START_LATITUDE.2 START_LATITUDE.1 BOTTOM_DEPTH.1 BOTTOM_DEPTH.2
  A     5829    49.0             45.0             183            549
  B     4024    49.0             45.0             549            900
  C     9259    49.0             40.5             900            1280
  D     6211    45.0             40.5             183            549
  E     5264    45.0             40.5             549            900
  F     6952    40.5             34.5             183            549
  G     7801    40.5             34.5             549            900
  H     8059    40.5             34.5             900            1280
}

SA3 is provided in this package.
}
\author{
Allan Hicks and Chantel Wetzel with help from John Wallace.
}
