% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataBiomass.fn.R
\name{strataBiomass.fn}
\alias{strataBiomass.fn}
\title{This function outputs a dataframe in the abundance index format needed for SS3
it calculates the standard error of log(B) by first finding the cv of B
It can output the median biomass or the mean biomass. Median biomass is bias corrected so that log(Bmedian)=mean(log(B))}
\usage{
strataBiomass.fn(bio, theStrata = "Year", fleet = "EnterFleet",
  season = 1, outputMedian = T)
}
\arguments{
\item{bio}{object}

\item{theStrata}{year}

\item{fleet}{Fleet number}

\item{season}{Season number}

\item{outputMedian}{TRUE/FALSE}
}
\description{
This function outputs a dataframe in the abundance index format needed for SS3
it calculates the standard error of log(B) by first finding the cv of B
It can output the median biomass or the mean biomass. Median biomass is bias corrected so that log(Bmedian)=mean(log(B))
}
\author{
Allan Hicks
}
