% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrataFactors.fn.R
\name{StrataFactors.fn}
\alias{StrataFactors.fn}
\title{Creates a vector of strata factors by name}
\usage{
StrataFactors.fn(dat, strat.vars, strat.df)
}
\arguments{
\item{strat.vars}{strata}

\item{strat.df}{strata data frame}

\item{data}{object}
}
\description{
Creates a vector of strata factors by name
}
\author{
Allan Hicks
}
