% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAFs.EWC.fn.R
\name{SurveyAFs.EWC.fn}
\alias{SurveyAFs.EWC.fn}
\title{Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datA should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS3, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SS3out: if True the output is in a format pastable into SS3 dat file}
\usage{
SurveyAFs.EWC.fn(dir, datA, datTows, strat.vars = NULL, strat.df = NULL,
  femaleMale = c(2, 1), ageBins = 1, SS3out = FALSE,
  meanRatioMethod = TRUE, gender = 3, NAs2zero = T,
  sexRatioUnsexed = NA, maxSizeUnsexed = NA, partition = 0,
  fleet = "Enter Fleet", agelow = "Enter", agehigh = "Enter",
  ageErr = "Enter", nSamps = "Enter Samps", season = "Enter Season",
  printfolder = "forSS")
}
\arguments{
\item{dir}{directory}

\item{datA}{object}

\item{strat.df}{}

\item{SS3out}{if True the output is in a format pastable into SS3 dat file}

\item{gender}{gender value for Stock Synthesis}

\item{NAs2zero}{change NAs to zeros}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{agelow}{value for SS -1}

\item{agehigh}{value for SS -1}

\item{ageErr}{age error vector to apply}

\item{nSamps}{effective sample size for Stock Synthesis}

\item{season}{season}

\item{printfolder}{folder where the length comps will be saved}

\item{lageBins}{length bins}
}
\description{
Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datA should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS3, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SS3out: if True the output is in a format pastable into SS3 dat file
}
\seealso{
\code{\link{SurveyLFs.EWC.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
