% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnexpandedAF.fn.R
\name{UnexpandedAFs.fn}
\alias{UnexpandedAFs.fn}
\title{Creates a matrix of length or age composition data WITHOUT expantion
Written by Chantel Wetzel to work with the data warehouse data formatting,}
\usage{
UnexpandedAFs.fn(
  dir = NULL,
  datA,
  ageBins = 1,
  sex = 3,
  partition = 0,
  fleet = "Enter Fleet",
  ageErr = "NA",
  agelow = -1,
  agehigh = -1,
  nSamps = "Enter Samps",
  month = "Enter Month",
  printfolder = "forSS",
  verbose = TRUE
)
}
\arguments{
\item{dir}{directory this is where the output files will be saved}

\item{datA}{the read in length comps by the PullBio.fn function}

\item{sex}{(0 = unsexed, 1 = females, 2 = males, 3 = females then males) sex value for Stock Synthesis}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{ageErr}{Number of ageing error matrix for SS}

\item{agelow}{age bin for SS (default value of -1)}

\item{agehigh}{age bin for SS (default value of -1)}

\item{nSamps}{effective sample size for Stock Synthesis}

\item{month}{month the samples were collected}

\item{printfolder}{folder where the length comps will be saved}

\item{verbose}{opt to print out message statements}

\item{lgthBins}{length bins}

\item{remove999}{the output object by the function will have the 999 column combined with the first length bin}
}
\description{
Creates a matrix of length or age composition data WITHOUT expantion
Written by Chantel Wetzel to work with the data warehouse data formatting,
}
\author{
Chantel Wetzel
}
