% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportion.R
\name{wh_plot_proportion}
\alias{wh_plot_proportion}
\title{Save figures of proportions by depth and latitude using warehouse data}
\usage{
wh_plot_proportion(
  data_catch,
  data_bio,
  dir = file.path(getwd(), "plots"),
  bar_width = c("n", "equal")
)
}
\arguments{
\item{data_catch, data_bio}{Data frames returned from \code{\link[=pull_catch]{pull_catch()}} and
[pull_bio())], respectively. At least one of the arguments must be passed.

[pull_bio())]: R:pull_bio())}

\item{dir}{The directory where you would like the \code{.png} files to be saved.
The default is a directory called \code{"plots"} in your current working
directory.}

\item{bar_width}{A string of \code{"n"} or \code{"equal"}, where the former leads to
bar widths based on the number of observations contained in that group and
the latter leads to equally-sized bars for all bars with data. For groups
without any data, the width of the placeholder on the x axis will be
smaller than the width of the bars for groups with data regardless of which
option you choose. The default is to have bars of variable widths, i.e.,
\code{"n"}.}
}
\value{
Strings of the saved files.
}
\description{
Four figures in total are created and saved to the disk if both catch and
biological data, pulled from the data warehouse, are passed to \code{data_catch} and
\code{data_bio}, respectively. This function will only work with data that has the
standard column names of data pulled from the warehouse.
}
\examples{
\dontrun{
#test <- wh_plot_proportion(catch_nwfsc_combo, bio_nwfsc_combo)
}
}
\seealso{
\itemize{
\item \code{\link[=plot_proportion]{plot_proportion()}}
\item \code{\link[purrr:map]{purrr::map()}}
}
}
\author{
Chantel R. Wetzel and Kelli F. Johnson
}
\concept{warehouse}
