% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomass.fn.R
\name{Biomass.fn}
\alias{Biomass.fn}
\title{Calculates design based estimates from survey data for West Coast surveys.
The design based index is calculated based on the area of the strata and
the mean catch by strata.
See: Gunderson, D.R. and Sample, T.M. 1980. Distribution and abundance of rockfish off Washington,
Oregon, and California during 1977. Marine Fisheries Review: March - April.}
\usage{
Biomass.fn(
  dir = NULL,
  dat,
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df,
  printfolder = "forSS",
  outputMedian = T,
  convert = 1,
  month = NA,
  fleet = NA,
  verbose = TRUE
)
}
\arguments{
\item{dir}{directory where the output file will be saved}

\item{dat}{data-frame of the data that has been by the PullCatch.fn}

\item{strat.vars}{A vector of the strata variable names (i.e., c("Depth_m","Latitude_dd"))}

\item{strat.df}{a dataframe with the first column the name of the stratum, the second column the area of the stratum, and the remaining columns are the high and low variables defining the strata created by the CreateStrataDF.fn}

\item{printfolder}{the folder where files will be saved}

\item{outputMedian}{T/F output median or the mean biomass estimate}

\item{convert}{convertion factor for the biomass by area units (e.g., hectare vs km2)}

\item{month}{month for SS}

\item{fleet}{fleet number for SS}

\item{verbose}{opt to print out message statements}
}
\description{
The variables defining the strata must begin with the name in strat.vars and end with ".1" or ".2" (i.e., Depth_m.1)
the strata are assumed to be continuous variables, thus have a lower and upper value defining them. The lower value does not necessarily have to be the same as the previous upper value.
the stat.df dataframe is difficult to build up with more than one variable becuase it turns into a design where you have to define all areas, thus repeat the variables for one (like a design)
}
\details{
I calculate the variance given stratified sampling theory
I work in normal space, then calculate the statistics if B is lognormal
This is the Mean Ratio Estimate
}
\author{
Allan Hicks and Chantel Wetzel
}
