% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{Directory check}
\usage{
check_dir(dir, verbose = TRUE)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Directory check
}
\details{
Check that
\enumerate{
\item The user knows that the data will not be saved if \code{dir = NULL}.
\item The directory exists if it can be created.
\item The function fails if the directory cannot be created.
}
}
\examples{
check_dir(getwd(), verbose = FALSE)
# See more output
check_dir(getwd())

}
\seealso{
Other helper function: 
\code{\link{check_survey}()},
\code{\link{combine_tows}()},
\code{\link{createMatrix}()},
\code{\link{filter_pull}()}
}
\author{
Chantel R. Wetzel
}
\concept{helper function}
