% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strata.R
\name{check_strata}
\alias{check_strata}
\title{Calculate the number of observations by year and strata}
\usage{
check_strata(data, strata, dir = NULL, printfolder = "forSS3", verbose = TRUE)
}
\arguments{
\item{data}{Data frame of catch data that has been created by the \code{\link[=pull_catch]{pull_catch()}}.}

\item{strata}{A data frame that defines the strata and provides the
calculated areas for each strata returned from \code{\link[=createStrataDF.fn]{createStrataDF.fn()}}.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{printfolder}{A string that will be appended to \code{dir}, creating a folder
where the output will be saved. If specified as \code{""},
the output will just be saved directly in \code{dir}. The default is \code{"forSS3"}.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
A matrix with the number of tows and positive tows within each strata
by year and the number of positive tows by strata and year.
}
\description{
Calculate the number of observations by year and strata
}
\details{
Calculates and returns the total number of tows and
positive tows conducted in each strata by year. The
selected stratas are used to expand the length and
marginal age compositions and to calculate a design
based index using the \code{\link[=get_design_based]{get_design_based()}}. In earlier
versions of the code, there needed to be more than one
positive observations within each strata to calculate
a design based index using \code{\link[=Biomass.fn]{Biomass.fn()}}. The new
\code{\link[=get_design_based]{get_design_based()}} function is more robust and will
return zeros in each strata-year combination with no
observations. However, it can be useful to review how
many tows and positive tows are present by year and
strata in your data to ensure that the selected strata
for expanding the data is reasonable (e.g., avoiding
limited observations in large areas).
}
\author{
Chantel Wetzel
}
