% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_survey.R
\name{check_survey}
\alias{check_survey}
\title{Check and create survey string}
\usage{
check_survey(survey)
}
\arguments{
\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.
}}
}
\description{
Check and create survey string
}
\author{
Chantel Wetzel
}
