% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStrata.fn.R
\name{CheckStrata.fn}
\alias{CheckStrata.fn}
\title{Calculate the number of observations by year and strata}
\usage{
CheckStrata.fn(
  dir = NULL,
  dat,
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df,
  printfolder = "forSS3",
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{Data-frame of the catch data that has been
created by \code{\link[=pull_catch]{pull_catch()}}.}

\item{strat.vars}{Variables in both data frame that are used to
define the strata. Default is bottom depth (m) and latitudes (decimal
degrees), i.e., \code{c("Depth_m", "Latitude_dd")}.}

\item{strat.df}{A data frame that defines the strata and provides the
calculated areas for each strata returned from \code{\link[=createStrataDF.fn]{createStrataDF.fn()}}.}

\item{printfolder}{A string that will be appended to \code{dir}, creating a folder
where the output will be saved. If specified as \code{""},
the output will just be saved directly in \code{dir}. The default is \code{"forSS3"}.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
A matrix with the number of tows within each strata by year and the
number of positive tows by strata and year.
}
\description{
Calculate the number of observations by year and strata
}
\details{
Calculates and returns the total number of tows and
positive tows conducted in each strata by year. The
selected strata are used to expand the length and
marginal age compositions and to calculate a design
based index using the \code{\link[=Biomass.fn]{Biomass.fn()}}.
}
\author{
Chantel Wetzel
}
