% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tows.R
\name{combine_tows}
\alias{combine_tows}
\title{Combine catch data by trawl_id for data pulled using the
pull_catch function. Generally, only a single tow is returned
by species and in those instances this function is not needed.
This function is only needed in unique data situations. For
example, if data are pulled with all sample_types included then
there will often be data from the same trawl_id included as
multiple records. In these instances this function allows for
the sample data to be combined for unique trawl_id.}
\usage{
combine_tows(dir = NULL, data, verbose = TRUE)
}
\arguments{
\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{data}{A data frame of catches obtained by using the pull_catch
function.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Combine catch data by trawl_id for data pulled using the
pull_catch function. Generally, only a single tow is returned
by species and in those instances this function is not needed.
This function is only needed in unique data situations. For
example, if data are pulled with all sample_types included then
there will often be data from the same trawl_id included as
multiple records. In these instances this function allows for
the sample data to be combined for unique trawl_id.
}
\author{
Chantel Wetzel
}
