% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMatrix.fn.R
\name{createMatrix}
\alias{createMatrix}
\title{Survey name matching function used when pull from the data warehouse
The website is: https://www.nwfsc.noaa.gov/data}
\usage{
createMatrix()
}
\description{
Survey name matching function used when pull from the data warehouse
The website is: https://www.nwfsc.noaa.gov/data
}
\author{
Chantel Wetzel based on code by John Wallace
}
