% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_vbgrowth.R
\name{est_vbgrowth}
\alias{est_vbgrowth}
\title{Estimate length-at-age using the von Bertanlaffy
parametization}
\usage{
est_vbgrowth(
  data,
  dir = NULL,
  col_length = "length_cm",
  col_age = "age",
  init_params = data.frame(K = 0.13, Linf = 55, L0 = 15, CV0 = 0.1, CV1 = 0.1),
  bins = NULL
)
}
\arguments{
\item{data}{Data frame of biological data from \code{\link[=pull_bio]{pull_bio()}}}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{col_length}{A numeric or character value specifying the column
to use in \code{data} for length information. These lengths are assumed to
be in centimeters. The default value is \code{length_cm}.}

\item{init_params}{Data frame of starting parameters for K, Linf, L0, CV0, and CV2 based on the
Stock Synthesis parameterization of von Bertanlaffy growth.}

\item{bins}{The bins to put ages into. If NULL then simply uses the ages as recorded.}

\item{col_weight}{A numeric or character value specifying the column
to use in \code{data} for weight information. These weights are assumed to
be in kilograms The default value is \code{weight_kg}.}
}
\description{
Estimate length-at-age using the von Bertanlaffy
parametization
}
\seealso{
\code{\link[=fit_vbgrowth]{fit_vbgrowth()}}
}
\author{
Chantel Wetzel
}
