% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format.AKSlope.fn.R
\name{Format.AKSlope.fn}
\alias{Format.AKSlope.fn}
\title{Rename columns in the AKFSC Slope data}
\usage{
Format.AKSlope.fn(dir = NULL, datTows, datL, start.year = 1997)
}
\arguments{
\item{dir}{directory}

\item{datTows}{input catch data frame}

\item{datL}{input biology list}

\item{start.year}{first year of data to include}
}
\description{
Rename columns in the AKFSC Slope data
}
\author{
Chantel Wetzel
}
