% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hkl_bio_data.R
\name{format_hkl_bio_data}
\alias{format_hkl_bio_data}
\title{Format NWFSC Hook-and-Line data}
\usage{
format_hkl_bio_data(data, common_name = NULL, verbose = TRUE)
}
\arguments{
\item{data}{Data frame of NWFSC Hook-and-Line data.}

\item{common_name}{Species names as given in the hook and line data set to
format data for.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
Returns a data frame of formatted hkl data
}
\description{
Format NWFSC Hook-and-Line data
}
\author{
Chantel Wetzel
}
