% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hkl_catch_data.R
\name{format_hkl_catch_data}
\alias{format_hkl_catch_data}
\title{Format NWFSC Hook-and-Line data}
\usage{
format_hkl_catch_data(
  data,
  common_name = NULL,
  return_positive_sites_only = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame of NWFSC Hook-and-Line data.}

\item{common_name}{Species names as given in the hook and line data set to
format data for.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}

\item{return_positive_site_only}{Logical that specifies whether data from sites
that have observed common_name at least once across the time series are
returned.The default is \code{TRUE}.}
}
\value{
Returns a data frame of formatted hkl data
}
\description{
Format NWFSC Hook-and-Line data
}
\author{
Chantel Wetzel
}
