% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_n.R
\name{get_input_n}
\alias{get_input_n}
\title{Calculate input sample sizes}
\usage{
get_input_n(
  dir = NULL,
  data,
  comp_column_name = c("length_cm", "age")[1],
  input_sample_size_method = c("stewart_hamel", "tows", "total_samples")[1],
  species_group = c("all", "flatfish", "shelfrock", "sloperock", "thorny", "other")[1],
  printfolder = "forSS3",
  verbose = TRUE
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{data}{A data frame of composition data created using \code{\link[=pull_bio]{pull_bio()}}.}

\item{comp_column_name}{The column name to create composition data for. This column can be
is used to determine whether to format the composition data for length or age
compositions by looking for either age (e.g., age_years, Age, best_age) or length
(e.g., Length, length, Length_cm) in the comp_column_name. Default Length_cm.}

\item{input_sample_size_method}{Determines the default input sample size to add to
the composition data for SS3. There are three options: c("stewart_hamel", "tows",
"total_samples") where the default is "stewart_hamel".}

\item{species_group}{A string specifying the species group of interest, which
will lead to the use of the correct species-specific value for
the number of unique samples per tow. See the function call for
allowed values, where the default is \code{"all"}.}

\item{printfolder}{A string that will be appended to \code{dir}, creating a folder
where the output will be saved. If specified as \code{""},
the output will just be saved directly in \code{dir}. The default is \code{"forSS3"}.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Calculate input sample sizes
}
\examples{
\dontrun{
bio <- pull_bio(
  common_name = "petrale sole",
  survey = "NWFSC.Combo"
)

n <- get_input_n(
  data = bio,
  comp_column_name = "length_cm",
  input_sample_size_method = "stewart_hamel",
  species_group = "flatfish"
)
}

}
\references{
Stewart, I.J. and O.S. Hamel. 2014.
Bootstrapping of sample size for length- or age-composition data used in
stock assessment.
Canadian Journal of Fishery and Aquatic Science, 71(4): 581--588.
\href{https://doi.org/10.1139/cjfas-2013-0289}{10.1139/cjfas-2013-0289}.
}
\author{
Chantel R. Wetzel
}
