% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetN.fn.R
\name{GetN.fn}
\alias{GetN.fn}
\title{Calculate effN input sample sizes
based on Stewart & Hamel 2014}
\usage{
GetN.fn(dir, dat, type, species = NULL, printfolder = "forSS")
}
\arguments{
\item{dir}{directory}

\item{dat}{excel file name for the NWFSC or data object for the Triennial or AKSFC survey data}

\item{type}{specify whether doing "length" or "age". Used to read associatied excel sheets}

\item{species}{species specific value to determine the number of unique samples per tow (flatfish, shelfrock, sloperock, thorny, others, all)}

\item{printfolder}{name of the folder to create and save files. Location will be paste0(dir, printfolder)}

\item{survey}{survey name, options = c("NWFSCBT", "Tri.Shelf", "AFSC.Slope")}

\item{sheet}{sheet name in excel file}

\item{headerRow}{line of header row in excel file}
}
\description{
Calculate effN input sample sizes
based on Stewart & Hamel 2014
}
\seealso{
\code{\link{readDataFromExcel.fn}}
}
\author{
Chantel Wetzel
}
