% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTotalBiomassExcel.fn.R
\name{GetTotalBiomassExcel.fn}
\alias{GetTotalBiomassExcel.fn}
\title{Wrapper function  to create the designed based indices}
\usage{
GetTotalBiomassExcel.fn(dir, file, sheet = "BiomassAbundance",
  printfolder = "forSS", headerRow, fleet = "EnterFleet", season = 1,
  outputMedian = T)
}
\arguments{
\item{dir}{directory of the excel file}

\item{file}{the excel file}

\item{sheet}{sheet name}

\item{printfolder}{the folder where files will be saved}

\item{headerRow}{header row in the csv file}

\item{fleet}{fleet number}

\item{season}{season number}

\item{outputMedian}{TRUE/FALSE}
}
\description{
Wrapper function  to create the designed based indices
}
\seealso{
\code{\link{readInExcelBiomass.fn}}, \code{\link{SS3Biomass.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
