% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_length_sampling.R
\name{plot_age_length_sampling}
\alias{plot_age_length_sampling}
\title{Plot the representativeness of age sampling based on lengths}
\usage{
plot_age_length_sampling(
  data,
  xlim = c(0, 120),
  ylim = c(0, 0.049),
  dir = NULL
)
}
\arguments{
\item{data}{data frame}

\item{xlim}{x limits for plot, defaults to c(0,120)}

\item{ylim}{y limits for plot, defaults to (0, 0.049)}

\item{dir}{Defaults to NULL (plot made, but not saved to file). Can alternatively be a string filename
by year and sex}
}
\description{
Plot the representativeness of age sampling based on lengths
}
