% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bio_patterns.R
\name{plot_bio_patterns}
\alias{plot_bio_patterns}
\title{This function plots length by latitude and depth}
\usage{
plot_bio_patterns(
  data,
  dir = NULL,
  col_name = "Length_cm",
  plot = 1:3,
  width = 7,
  height = 7,
  verbose = TRUE
)
}
\arguments{
\item{data}{Biological data frame from \code{\link[=pull_bio]{pull_bio()}}}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{col_name}{Option to switch between plotting lengths or ages.
Options are "Length_cm", "Width_cm", or "Age".}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print or save both figures, i.e., \code{plot = 1:3}. Integers
correspond to the following figures:
\enumerate{
\item length/age by latitude and depth
\item length/age by depth and year
\item length/age by lat and year
}}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 10 by 7 inches.}
}
\description{
This function plots length by latitude and depth
}
\seealso{
Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_var_length_at_age}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Chantel Wetzel
}
\concept{plot_}
