% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bio_patterns.R
\name{plot_bio_patterns}
\alias{plot_bio_patterns}
\title{This function plots length by latitude and depth}
\usage{
plot_bio_patterns(
  dir = NULL,
  bio,
  col_name = "Length_cm",
  plot = 1:3,
  width = 7,
  height = 7
)
}
\arguments{
\item{dir}{Directory to save files to}

\item{bio}{Data biological sample file}

\item{col_name}{Option to switch between plotting lengths or ages.
Options are "Length_cm", "Width_cm", or "Age".}

\item{width}{Numeric figure width in inches, defaults to 7}

\item{height}{Numeric figure height in inches, defaults to 7}
}
\description{
This function plots length by latitude and depth
}
\author{
Chantel Wetzel
}
