% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{This function plots cpue and length by latitude and depth}
\usage{
plot_cpue(catch, dir = NULL, plot = 1:3, ...)
}
\arguments{
\item{catch}{Data catch file pulled using \code{\link[=pull_catch]{pull_catch()}}}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{plot}{A vector of integers specifying the figures you want.}

\item{...}{additional arguments to \code{\link[=ggsave]{ggsave()}}. Figure width and height default to 7 in.}
}
\description{
This function plots cpue and length by latitude and depth
}
\author{
Chantel Wetzel
}
