% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{This function plots cpue and length by latitude and depth}
\usage{
plot_cpue(dir = NULL, catch, plot = 1:3, width = 7, height = 7)
}
\arguments{
\item{dir}{Directory to save files to}

\item{catch}{Data catch file}

\item{plot}{A vector of integers specifying the figures you want.}

\item{width}{Numeric figure width in inches, defaults to 7}

\item{height}{Numeric figure height in inches, defaults to 7}
}
\description{
This function plots cpue and length by latitude and depth
}
\author{
Chantel Wetzel
}
