% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue_map.R
\name{plot_cpue_map}
\alias{plot_cpue_map}
\title{Plot two figures showing catch-per-unit-effort data}
\usage{
plot_cpue_map(
  data,
  dir = NULL,
  main = NULL,
  plot = 1:2,
  height = 10,
  width = 7
)
}
\arguments{
\item{data}{A dataframe of catch data created by \code{\link[=pull_catch]{pull_catch()}}.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{main}{A string that will be prepended to the name of the saved png
(i.e., "NWFSC" results in a file called "NWFSC_CPUE_Map.png").}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print both figures.
\enumerate{
\item coastwide data across all years
\item coastwide data by year
}}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 10 inches.}
}
\value{
Figures are saved to the disk according to which plots are asked
for in \code{plot}. Each of the specified files are saved to \code{dir}, the specified
directory. No objects are returned to the user. But, the figures are printed
to new windows if they are not saved to the disk.
}
\description{
Plot catch-per-unit-effort (CPUE) data across all years and by year
and save them to the disk. Figures are created using {ggplot2}.
}
\examples{
\dontrun{
plot_cpue_map(data = catch_nwfsc_combo, plot = 1)
plot_cpue_map(data = catch_nwfsc_combo, plot = 2)
}

}
\seealso{
\itemize{
\item \code{pull_catch()}
\item \code{plot_westcoast()}
}

Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_bio_patterns}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_var_length_at_age}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Chantel R. Wetzel
}
\concept{plot_}
