% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportion.R
\name{plot_proportion}
\alias{plot_proportion}
\title{Plot showing presence and absence per haul by depth or latitute bin}
\usage{
plot_proportion(
  data,
  dim = c("depth", "lat", "sex")[1],
  dir = NULL,
  plot_type = c("proportion", "total")[1],
  depth_min = 50,
  depth_max = NULL,
  depth_bin_width = 25,
  lat_min = 32,
  lat_max = 49,
  lat_bin_width = 1,
  add_range_to_main = TRUE,
  xlab = NULL
)
}
\arguments{
\item{data}{data.frame containing data per haul created by
PullCatch.fn() or biological data created by PullBio.fn() where the
dim input must be "sex".}

\item{dim}{Dimension of interest, either "depth", "lat", or "sex".}

\item{dir}{The directory name to save figures to where they will be save
via file.path(dir, "plots").}

\item{plot_type}{Two options area avialable "proportion" or "total" where
the default, "proportion", plot the proportion by depth/latitude with equal
bar widths and the "total" option plot the numbers by depth/latitude were
the bar widths in a production of sampling by depth/latitude.}

\item{depth_min}{Minimum depth (in meters)}

\item{depth_max}{Maximum depth (in meters). A NULL value will cause the
function to automatically set depth_max to the multiple of depth_bin_width beyond
the 99.9 percentile of the observations.}

\item{depth_bin_width}{Width of each depth bin (in meters)}

\item{lat_min}{Minimum latitute (in decimal degrees)}

\item{lat_max}{Maximum latitute (in decimal degrees)}

\item{lat_bin_width}{Width of each latitude bin (in decimal degrees)}

\item{add_range_to_main}{Add the range of latitude or depth by which the data
are filtered}

\item{xlab}{Label for x-axis. A NULL value will cause the function
to choose either "Depth (m)" or "Latitude (°N)"}
}
\description{
Uses the data.frame of data values extracted from the data warehouse
by either \code{\link[=PullCatch.fn]{PullCatch.fn()}} or \code{\link[=PullBio.fn]{PullBio.fn()}} to make a table of
proportion by depth or latitude bins. If biological data are passed
to the function then there is a single option to plot the sex
ratio by depth.
}
\examples{
# load WCGBTS data data
data.WCGBTS.ling <- nwfscSurvey::PullCatch.fn(
  Name = "lingcod",
  SurveyName = "NWFSC.Combo"
)
bio.WCGBTS.ling <- nwfscSurvey::PullBio.fn(
  Name = "lingcod",
  SurveyName = "NWFSC.Combo"
)
plot_proportion(data = data.WCGBTS.ling, dim = "lat")
plot_proportion(data = data.WCGBTS.ling, dim = "depth")
plot_proportion(data = bio.WCGBTS.ling, dim = "sex")

}
\author{
Ian G. Taylor and Chantel Wetzel
}
