% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sex_ratio.R
\name{plot_sex_ratio}
\alias{plot_sex_ratio}
\title{Function to plot sex ratio data in a barplot form}
\usage{
plot_sex_ratio(
  data,
  dir = NULL,
  comp_column_name = "length_cm",
  main = NULL,
  bin_width = ifelse(comp_column_name == "length_cm", 2, 1),
  width = 7,
  height = 7
)
}
\arguments{
\item{data}{Data object with biological data from \code{\link[=pull_bio]{pull_bio()}} with a column
named \code{Sex} or \code{sex} is present.}

\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{comp_column_name}{The name of the column to create composition data for that
must be a string. This column can be is used to determine whether to format the
composition data for length or age compositions by looking for either age
(e.g., \code{age_years}, \code{Age}, \code{age}, \code{best_age}) or length (e.g., \code{length},
\code{length_cm}, \code{Length}, \code{Length_cm}) in the comp_column_name. The comp_column_name
is not case sensitive.The default is \code{length_cm}.}

\item{main}{Unique string that will be added to the saved png name.  This can
be useful when plotting data for various data sets separately.}

\item{bin_width}{Width to bin the data by.  The default is 2 cm for lengths
and 1 for ages.}

\item{width, height}{Numeric values for the figure width and height in
inches. The defaults are 7 by 7 inches.}
}
\description{
Function to plot sex ratio data in a barplot form
}
\seealso{
Other plot_: 
\code{\link{PlotSexRatio.fn}()},
\code{\link{PlotVarLengthAtAge.fn}()},
\code{\link{plot_age_length_sampling}()},
\code{\link{plot_comps}()},
\code{\link{plot_cpue}()},
\code{\link{plot_cpue_map}()},
\code{\link{plot_index}()},
\code{\link{plot_length_age}()},
\code{\link{plot_proportion}()},
\code{\link{plot_sex_ratio_strata}()},
\code{\link{plot_var_length_at_age}()},
\code{\link{plot_weight_length}()},
\code{\link{plot_westcoast}()}
}
\author{
Chantel Wetzel
}
\concept{plot_}
