% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_westcoast.R
\name{plot_westcoast}
\alias{plot_westcoast}
\title{Plot the United States west coast using {ggplot2}}
\usage{
plot_westcoast()
}
\value{
A {ggplot2} object.
}
\description{
Plot the United States west coast such that it can used as a standalone
figure or as a base to plot data points on.
}
\examples{
\dontrun{
if (require("maps") && require("mapproj")) {
map_object <- plot_westcoast()
print(map_object)
ggplot2::ggsave(
  filename = "myfilename.png",
  plot =  map_object
)
}
}
}
\seealso{
\itemize{
\item \code{PlotMap.fn()} uses this function as the base map.
}
}
\author{
Kelli F. Johnson
}
