% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBio.fn.R
\name{PlotBio.fn}
\alias{PlotBio.fn}
\title{Plots the designed based biomass estimates with confidence intervals}
\usage{
PlotBio.fn(
  dir = NULL,
  dat,
  CI = 0.95,
  scalar = 1e+06,
  gap = 0.03,
  ylab = "Biomass ('000 mt)",
  xlab = "Year",
  main = NULL,
  ylim = NULL,
  add = FALSE,
  col = "black",
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{object created by the \code{\link[=Biomasss.fn]{Biomasss.fn()}}}

\item{CI}{A numerical value that specifies the confidence interval to return.
Values should be between 0.01 to 0.99.}

\item{scalar}{simply the divisor for the biomass}

\item{gap}{a value that introduces a slight gap between the point estimate and the start of the line for the CI. A gap too large will invert the CI, making it look huge. You should know when this happens}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{main}{plot label}

\item{ylim}{y-limits}

\item{add}{add additional line to plot}

\item{col}{color}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots}
}
\description{
Plots the designed based biomass estimates with confidence intervals
}
\author{
Chantel Wetzel, Allan Hicks, and John Wallace
}
