% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioStrata.fn.R
\name{plotBioStrata.fn}
\alias{plotBioStrata.fn}
\title{Not recommended for use yet
Name would imply that it would plot by strata but that is currently not happening
Duplicate to the plotBio.fn but not as polished}
\usage{
plotBioStrata.fn(bio, CI = 0.95, scalar = 1e+06, gap = 0.03,
  ylab = "Biomass ('000 mt)", xlab = "Stratum", ylim = NULL, ...)
}
\arguments{
\item{bio}{object created by the GetTotalBiomass.fn}

\item{CI}{confidence interval}

\item{scalar}{simply the divisor for the biomass}

\item{gap}{a value that introduces a slight gap between the point estimate and the start of the line for the CI. A gap too large will invert the CI, making it look huge. You should know when this happens}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{ylim}{y-limits}
}
\description{
Not recommended for use yet
Name would imply that it would plot by strata but that is currently not happening
Duplicate to the plotBio.fn but not as polished
}
\author{
Allan Hicks
}
