% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioStrata.fn.R
\name{plotBioStrata.fn}
\alias{plotBioStrata.fn}
\title{Plot the design based estimated biomass by strata and year
Currently only works with the triennial or afsc survey data
This is a new function and the results should be scrutinized}
\usage{
plotBioStrata.fn(dir, dat, CI = 0.95, scalar = 1e+06, gap = 0.01,
  legloc = "topright", single = TRUE, ylab = "Biomass ('000 mt)",
  xlab = "Year", yliminput = NULL, dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory to save the file}

\item{dat}{object created by the GetTotalBiomass.fn}

\item{CI}{confidence interval}

\item{scalar}{simply the divisor for the biomass}

\item{gap}{a value that introduces a slight gap between the point estimate and the start of the line for the CI. A gap too large will invert the CI, making it look huge. You should know when this happens}

\item{legloc}{location to put legend in the single panel plot}

\item{single}{TRUE/FALSE true plots each stratum on the same figure}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{yliminput}{user defined y-limits}

\item{dopng}{TRUE/FALSE whether to save a png file}
}
\description{
Plot the design based estimated biomass by strata and year
Currently only works with the triennial or afsc survey data
This is a new function and the results should be scrutinized
}
\author{
Allan Hicks and Chantel Wetzel
}
