% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreqData.fn.R
\name{plotFreqData.fn}
\alias{plotFreqData.fn}
\title{This function plots frequency data as bubble plots
You may want to change all zeros to NA's so that those observations are not plotted.
If you don't then set zero2NAs=F}
\usage{
plotFreqData.fn(dir, dat, survey = "Survey", inch = 0.15, ylab = "Bins",
  xlab = "Year", zero2NAs = T, main = NULL, xlim = NULL,
  dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory to save files to}

\item{dat}{object created by SS3LF.fn or SS3AF.fn}

\item{survey}{survey name}

\item{inch}{input to the symbols plot: TRUE, FALSE or a positive number.}

\item{ylab}{y-axis text label}

\item{xlab}{x-axis text label}

\item{zero2NAs}{T/F change 0 values to NA}

\item{main}{main plot text}

\item{xlim}{x-limit values}

\item{dopng}{save the plot as a png inside plots folder}
}
\description{
This function plots frequency data as bubble plots
You may want to change all zeros to NA's so that those observations are not plotted.
If you don't then set zero2NAs=F
}
\author{
Allan Hicks and Chantel Wetzel
}
