% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMap.fn.R
\name{PlotMap.fn}
\alias{PlotMap.fn}
\title{Plot two figures showing catch-per-unit-effort data}
\usage{
PlotMap.fn(
  dir = NULL,
  dat,
  main = NULL,
  dopng = lifecycle::deprecated(),
  plot = 1:2
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{An object created by \code{\link[=pull_catch]{pull_catch()}}.}

\item{main}{A string that will be prepended to the name of the saved png
(i.e., "NWFSC" results in a file called "NWFSC_CPUE_Map.png").}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{plot}{A vector of integers to specify which plots you would like. The
default is to print both figures.
\enumerate{
\item coastwide data across all years
\item coastwide data by year
}}
}
\value{
Figures are saved to the disk according to which plots are asked
for in \code{plot}. Each of the specified files are saved to a directory called
\code{map_plots} inside of \code{dir}, the specified directory. No objects are
returned to the user. But, the figures are printed to new windows if they
are not saved to the disk.
}
\description{
Plot catch-per-unit-effort (CPUE) data across all years and by year
and save them to the disk. Figures are created using {ggplot2}.
}
\examples{
\dontrun{
PlotMap.fn(dat = catch_nwfsc_combo, plot = 1)
PlotMap.fn(dat = catch_nwfsc_combo, plot = 2)
}
}
\seealso{
\itemize{
\item \code{PullCatch.fn()}
\item \code{plot_westcoast()}
}
}
\author{
Chantel R. Wetzel
}
