% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMap.fn.R
\name{PlotMap.fn}
\alias{PlotMap.fn}
\title{#This function will create two plots
\enumerate{
\item CPUE from survey data across all years, and
\item CPUE plot of the survey data by year.
}}
\usage{
PlotMap.fn(dir = NULL, dat, main = NULL, dopng = FALSE, plot = 1:2)
}
\arguments{
\item{dir}{directory}

\item{dat}{object created by the PullCatch.fn function}

\item{main}{name that will be used to name the saved png (i.e., "NWFSC" results in a file called "NWFSC_CPUE_Map.png")}

\item{dopng}{save the plot as a png inside plots folder}

\item{plot}{select a subset of plots: 1 = coastwide across all years, 2 = coastwide by year}
}
\description{
#This function will create two plots
\enumerate{
\item CPUE from survey data across all years, and
\item CPUE plot of the survey data by year.
}
}
\author{
Chantel Wetzel
}
