% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPresenceAbsence.fn.R
\name{PlotPresenceAbsence.fn}
\alias{PlotPresenceAbsence.fn}
\title{Plot showing presence and absence per haul by depth, latitude, or sex bin}
\usage{
PlotPresenceAbsence.fn(...)
}
\arguments{
\item{...}{Any arguments associated with now-deprecated function}
}
\description{
This function was deprecated in {nwfsc} version 2.1. Please use
\code{\link[=plot_proportion]{plot_proportion()}} instead.
}
