% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPresenceAbsence.fn.R
\name{PlotPresenceAbsence.fn}
\alias{PlotPresenceAbsence.fn}
\title{Plot showing presence and absence per haul by depth or latitute bin}
\usage{
PlotPresenceAbsence.fn(
  catch,
  dim = "depth",
  dir = NULL,
  dopng = FALSE,
  depth_min = 50,
  depth_max = NULL,
  depth_bin_width = 25,
  lat_min = 32,
  lat_max = 49,
  lat_bin_width = 1,
  main = "Presence/absence in WCGBT Survey",
  add_range_to_main = TRUE,
  xlab = NULL
)
}
\arguments{
\item{catch}{data.frame containing catch per haul created by
PullCatch.fn()}

\item{dim}{Dimension of interest, either "depth" or "lat"}

\item{dir}{The directory name to print to}

\item{dopng}{Boolean, whether or not to do png. Defaults to FALSE,}

\item{depth_min}{Minimum depth (in meters)}

\item{depth_max}{Maximum depth (in meters). A NULL value will cause the
function to automatically set depth_max to the multiple of depth_bin_width beyond
the 99.9 percentile of the observations.}

\item{depth_bin_width}{Width of each depth bin (in meters)}

\item{lat_min}{Minimum latitute (in decimal degrees)}

\item{lat_max}{Maximum latitute (in decimal degrees)}

\item{lat_bin_width}{Width of each latitude bin (in decimal degrees)}

\item{main}{String title, defaults to "Presence/absence in WCGBT Survey"}

\item{add_range_to_main}{Add the range of latitude or depth by which the data
are filtered}

\item{xlab}{Label for x-axis. A NULL value will cause the function
to choose either "Depth (m)" or "Latitude (°N)"}
}
\description{
Uses the data.frame of catch values extracted from the data warehouse
by [PullCatch.fn()] to make a table of zero and non-zero catch by
depth or latitute bins.
}
\examples{
# load WCGBTS catch data
catch.WCGBTS.ling <- nwfscSurvey::PullCatch.fn(
  Name = "lingcod",
  SurveyName = "NWFSC.Combo"
)
PlotPresenceAbsence.fn(catch.WCGBTS.ling, dim = "lat")
PlotPresenceAbsence.fn(catch.WCGBTS.ling, dim = "depth")
}
\author{
Ian G. Taylor
}
