% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSexRatio.fn.R
\name{PlotSexRatio.fn}
\alias{PlotSexRatio.fn}
\title{Function to plot sex ratio}
\usage{
PlotSexRatio.fn(dir, dat, type = "length", survey, circleSize = 0.1,
  dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory location for saving the png}

\item{dat}{data object}

\item{survey}{specify the survey data set for calculations. Options are "NWFSCBT", "Tri.Shelf", "AFSC.Slope"}

\item{circleSize}{circle size}

\item{dopng}{TRUE/FALSE whether to save a png file}

\item{...}{Additional arguments for the plots}
}
\description{
Function to plot sex ratio
}
\author{
Allan Hicks and Chantel Wetzel
}
