% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexRatio.fn.R
\name{PlotSexRatio.fn}
\alias{PlotSexRatio.fn}
\title{Function to plot sex ratio}
\usage{
PlotSexRatio.fn(
  dir,
  dat,
  data.type = "length",
  main = NULL,
  circleSize = 0.1,
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{Data object with biological data from \code{\link[=pull_bio]{pull_bio()}} with a column
named \code{Sex} present.}

\item{data.type}{Specify where to calculate the sex ration by length or age.}

\item{main}{Name that will be used to name the saved png}

\item{circleSize}{circle size}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots}
}
\description{
Function to plot sex ratio
}
\author{
Allan Hicks and Chantel Wetzel
}
