% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSexRatio.fn.R
\name{PlotSexRatio.fn}
\alias{PlotSexRatio.fn}
\title{Function to plot sex ratio}
\usage{
PlotSexRatio.fn(
  dir,
  dat,
  data.type = "length",
  main = NULL,
  circleSize = 0.1,
  dopng = FALSE,
  ...
)
}
\arguments{
\item{dir}{directory location for saving the png}

\item{dat}{data object}

\item{data.type}{"length" or "age"}

\item{main}{ame that will be used to name the saved png}

\item{circleSize}{circle size}

\item{dopng}{TRUE/FALSE whether to save a png file}

\item{...}{Additional arguments for the plots}
}
\description{
Function to plot sex ratio
}
\author{
Allan Hicks and Chantel Wetzel
}
