% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSexRatioStrata.fn.R
\name{PlotSexRatioStrata.fn}
\alias{PlotSexRatioStrata.fn}
\title{Function to plot sex ratio by strata}
\usage{
PlotSexRatioStrata.fn(
  dir = NULL,
  dat,
  type = "length",
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df = NULL,
  circleSize = 0.05,
  dopng = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{dir}{Directory where output will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{dat}{A data frame of length-composition data returned from
\code{\link[=pull_bio]{pull_bio()}}.}

\item{type}{length/age which data type to use}

\item{strat.vars}{Variables in both data frame that are used to
define the stratas. Default is bottom depth (m) and latitudes (decimal
degrees), i.e., \code{c("Depth_m", "Latitude_dd")}.}

\item{strat.df}{A data frame that defines the strata and provides the
calculated areas for each strata returned from \code{\link[=createStrataDF.fn]{createStrataDF.fn()}}.}

\item{circleSize}{circle size}

\item{dopng}{Deprecated with {nwfscSurvey} 2.1 because providing a non-NULL
value to \code{dir} can serve the same purpose as \code{dopng = TRUE} without the
potential for errors when \code{dopng = TRUE} and \code{dir = NULL}. Thus, users
no longer have to specify \code{dopng} to save the plot as a png.}

\item{...}{Additional arguments for the plots}
}
\description{
Function to plot sex ratio by strata
}
\seealso{
\code{\link{StrataFactors.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
