% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSexRatioStrata.fn.R
\name{PlotSexRatioStrata.fn}
\alias{PlotSexRatioStrata.fn}
\title{Function to plot sex ratio by strata}
\usage{
PlotSexRatioStrata.fn(dir, dat, type = "length", strat.vars = c("Depth_m",
  "Latitude_dd"), survey, circleSize = 0.05, dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory location for saving the png}

\item{dat}{data object}

\item{type}{length/age which data type to use}

\item{strat.vars}{the parameters to stratify the data}

\item{survey}{specify the survey data set for calculations. Options are "NWFSCBT", "Tri.Shelf", "AFSC.Slope"}

\item{circleSize}{circle size}

\item{dopng}{TRUE/FALSE whether to save a png file}

\item{...}{Additional arguments for the plots}
}
\description{
Function to plot sex ratio by strata
}
\seealso{
\code{\link{StrataFactors.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
