% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVarLengthAtAge.fn.R
\name{PlotvarLengthAtAge.fn}
\alias{PlotvarLengthAtAge.fn}
\title{Plot variability of length at age}
\usage{
PlotvarLengthAtAge.fn(dir, dat, survey, ageBin = 1, bySex = T,
  parStart = c(52, 0.09, 1), estVB = T, bins = NULL,
  legX = "bottomleft", legY = NULL, dopng = FALSE, ...)
}
\arguments{
\item{dir}{directory to save the file}

\item{dat}{The data loaded from the NWFSC database or the SexedLgthWtAge sheet}

\item{survey}{specify the survey data set for calculations. Options are "NWFSCBT", "Tri.Shelf", "AFSC.Slope"}

\item{ageBin}{Currently fixed at 1, so a moot parameter}

\item{bySex}{Logical to indicate if plot by sex}

\item{parStart}{Vector of starting parameters for Linf, k, and t0 in VonB estimation}

\item{estVB}{Logical. Estimate vonB growth to plot against predicted length. If F,
it uses the paramters in \code{parStart}.}

\item{bins}{The bins to put ages into. If NULL then simply uses the ages as recorded.}

\item{dopng}{TRUE/FALSE whether to save a png file}

\item{...}{Additional arguments for the plots}
}
\description{
Plot variability of length at age
}
\details{
Plots the SD and CV of age at observed and predicted length
}
\examples{
  slwa <- read.csv(paste(directory,"SexedLgthWtAge.csv",sep="\\\\"),skip=8)
  ageBin currently is fixed at 1 no matter what number you enter
  res <- varLengthAtAge.fn(slwa,ageBin=1,bySex=T,parStart=c(52,0.09,1),estVB=T)
}
\author{
Allan Hicks and Chantel Wetzel
}
