% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_catch.R
\name{pull_catch}
\alias{pull_catch}
\title{Pull catch data for satisfactory tows from the NWFSC data warehouse}
\usage{
pull_catch(
  common_name = NULL,
  sci_name = NULL,
  years = c(1980, 2050),
  survey = NULL,
  dir = NULL,
  convert = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{common_name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{sci_name} argument if you know the latin name.}

\item{sci_name}{A character entry with the desired scientific name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{common_name} argument if you know the common name.}

\item{years}{An integer vector of length two with the
range of years to pull data for.}

\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for. The input options are:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
The National Marine Fishery Service Alaska Fisheries Science Center (AFSC)
Triennial survey was conducted between 1977 - 2004 occurring every 3rd year.
The initial year, 1977, survey is not traditionally used in calculating
The AFSC Slope Survey (AFSC.Slope) along the west coast of the U.S. began in 1984 and occurred
annually from 1988–2001, with the exception of 1994 and 1998, when surveys were not conducted.
Prior to 1997, only a limited portion of the coast was covered in each year.
U.S. West Coast groundfish stock assessments only use the four years of consistent
and complete survey coverage (1997, 1999-2001). The Northwest Fisheries Science
Center (NWFSC) Slope survey (NWFSC.Slope) was conducted between 1998 - 2001.
The NWFSC West Coast Groundfish Bottom Trawl survey (NWFSC.Combo) is conducted
annualy starting in 2003 (excluding 2020) and samples both the U.S. west coast
shelf and slope between 55 - 1,280 meters.
Data can only be pulled from one survey at a time, though we are working on
allowing for a vector of survey names.
Currently, \code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.
}}

\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{convert}{TRUE/FALSE to convert column names to first letter uppercase
which aligns with the expected names in data processing functions.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Pull catch data from the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
for a single species or all observed species, where the latter is specified
by leaving both \code{common_name = NULL} and \code{sci_name = NULL}.
}
\details{
The data available in the warehouse are cleaned pior to being downloaded
with the intent that they provide the best available information for use
in an index-standardization procedure. The removed samples may be of use
to others with a less-restrictive goal than producing an index of abundance.
For example, life-stage samples are excluded because they are not collected
using the same protocols as standard samples.
To download all data, we currently recommend going to the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
and using the csv link to extract data for a single species at a time.
In the future, we hope to add functionality to this package such that
downloading all data can be done easily within this function.
See \href{https://github.com/pfmc-assessments/nwfscSurvey/issues/43}{Issue #43}
for more information.
}
\examples{
\dontrun{
# survey is only arg that has to be specified
catch_data <- pull_catch(survey = "NWFSC.Combo")

# Example with specified common name
catch_data <- pull_catch(common_name = "vermilion rockfish",
                 survey = "NWFSC.Combo")

# Example with specified scientific name
catch_data <- pull_catch(sci_name = "Eopsetta jordani",
                 survey = "NWFSC.Combo")

# Example with multiple names
catch_data <- pull_catch(common_name = c("vermilion rockfish",
                 "vermilion and sunset rockfish"), survey = "NWFSC.Combo")

catch_data <- pull_catch(sci_name = c("Sebastes miniatus",
                 "Sebastes sp. (crocotulus)",
                 "Sebastes sp. (miniatus / crocotulus)"),
                 survey = "NWFSC.Combo")
}

}
\author{
Chantel Wetzel
}
