% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_catch.R
\name{pull_catch}
\alias{pull_catch}
\title{Pull catch data for satisfactory tows from the NWFSC data warehouse}
\usage{
pull_catch(
  common_name = NULL,
  sci_name = NULL,
  years = c(1980, 2050),
  survey = NULL,
  dir = NULL,
  convert = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{common_name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{sci_name} argument if you know the latin name.}

\item{sci_name}{A character entry with the desired scientific name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{common_name} argument if you know the common name.}

\item{years}{An integer vector of length two with the
range of years to pull data for.}

\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.
}}

\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{convert}{TRUE/FALSE to convert column names to first letter uppercase
which aligns with the expected names in data processing functions.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Pull catch data from the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
for a single species or all observed species, where the latter is specified
by leaving both \code{common_name = NULL} and \code{sci_name = NULL}.
}
\details{
The data available in the warehouse are cleaned pior to being downloaded
with the intent that they provide the best available information for use
in an index-standardization procedure. The removed samples may be of use
to others with a less-restrictive goal than producing an index of abundance.
For example, life-stage samples are excluded because they are not collected
using the same protocols as standard samples.
To download all data, we currently recommend going to the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
and using the csv link to extract data for a single species at a time.
In the future, we hope to add functionality to this package such that
downloading all data can be done easily within this function.
See \href{https://github.com/nwfsc-assess/nwfscSurvey/issues/43}{Issue #43}
for more information.
}
\examples{
\dontrun{
# survey is only arg that has to be specified
catch_data <- pull_catch(survey = "NWFSC.Combo")

# Example with specified common name
catch_data <- pull_catch(common_name = "vermilion rockfish",
                 survey = "NWFSC.Combo")

# Example with specified scientific name
catch_data <- pull_catch(sci_name = "Eopsetta jordani",
                 survey = "NWFSC.Combo")

# Example with multiple names
catch_data <- pull_catch(common_name = c("vermilion rockfish",
                 "vermilion and sunset rockfish"), survey = "NWFSC.Combo")

catch_data <- pull_catch(sci_name = c("Sebastes miniatus",
                 "Sebastes sp. (crocotulus)", 
                 "Sebastes sp. (miniatus / crocotulus)"),
                 survey = "NWFSC.Combo")
}

}
\author{
Chantel Wetzel
}
