% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_gemm.R
\name{pull_gemm}
\alias{pull_gemm}
\title{Pull gemm data from the NWFSC data warehouse
The website is: https://www.webapps.nwfsc.noaa.gov/data
This function can be used to pull all gemm data, a single species, or a 
subset of species c("Canary Rockfish", "Widow Rockfish"). Species names in 
the gemm are capitalized (e.g. Canary Rockfish). However, there are checks in 
the function to adjust input species names if the input names do not match
the expected capitalization (e.g. "canary rockfish", "canary_rockfish"). The 
fuction also allows you to subset the data by year using the years input and to 
save the object if the dir function input is given.}
\usage{
pull_gemm(dir, species, years)
}
\arguments{
\item{dir}{directory to save file, if left blank pull will not be saved}

\item{species}{common name of species data to pull from the GEMM. The GEMM uses common
names with the first and last parts capitalized: Canary Rockfish}

\item{years}{range of years to pull data}
}
\description{
Pull gemm data from the NWFSC data warehouse
The website is: https://www.webapps.nwfsc.noaa.gov/data
This function can be used to pull all gemm data, a single species, or a 
subset of species c("Canary Rockfish", "Widow Rockfish"). Species names in 
the gemm are capitalized (e.g. Canary Rockfish). However, there are checks in 
the function to adjust input species names if the input names do not match
the expected capitalization (e.g. "canary rockfish", "canary_rockfish"). The 
fuction also allows you to subset the data by year using the years input and to 
save the object if the dir function input is given.
}
\examples{
\dontrun{
# Pull all GEMM data
all_data <- pull_gemm()

# Pull for a specific specis 
widow_data <- pull_gemm(species = "Widow Rockfish")

# Pull multiple species
data <- pull_gemm(species = c("Widow Rockfish", "Canary Rockfish"))

# Pull species and subset years
widow_recent <- pull_gemm(species = "Widow Rockfish", years = 2015:2019)
}


}
\author{
Chantel Wetzel
}
