% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_haul.R
\name{pull_haul}
\alias{pull_haul}
\title{Pull haul data from the NWFSC data warehouse.
The website is: https://www.webapps.nwfsc.noaa.gov/data.
This function can be used to pull haul data and associated covariates.}
\usage{
pull_haul(years = c(1980, 2050), survey = NULL, dir = NULL, verbose = TRUE)
}
\arguments{
\item{years}{An integer vector of length two with the
range of years to pull data for.}

\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.
}}

\item{dir}{directory where ouptut will be saved. The directory where the file should be saved.
If dir = NULL no output will be saved.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\value{
Returns a data frame of haul characteristics for satisfactory hauls
}
\description{
Pull haul data from the NWFSC data warehouse.
The website is: https://www.webapps.nwfsc.noaa.gov/data.
This function can be used to pull haul data and associated covariates.
}
\examples{
\dontrun{
haul_data <- pull_haul(survey = "NWFSC.Combo", years = c(2003, 2007))
haul_data <- pull_haul()
}

}
\author{
Eric Ward, Chantel Wetzel
}
