% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_spp.R
\name{pull_spp}
\alias{pull_spp}
\title{Pull species names from the warehouse}
\usage{
pull_spp()
}
\description{
Pull common name and scientific name information from the
data warehouse.
The website is https://www.webapps.nwfsc.noaa.gov/data
}
\examples{
\dontrun{
spp <- pull_spp()
}

}
\author{
Kelli Faye Johnson
}
