% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.fn.R
\name{PullCatch.fn}
\alias{PullCatch.fn}
\title{Pull catch data from the NWFSC data warehouse}
\usage{
PullCatch.fn(
  Name = NULL,
  SciName = NULL,
  YearRange = c(1980, 5000),
  SurveyName = NULL,
  SaveFile = FALSE,
  Dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Name}{common name of species data to pull from the data warehouse}

\item{SciName}{scientific name of species data to pull from the data warehouse}

\item{YearRange}{range of years to pull data}

\item{SurveyName}{survey to pull the data for the options are:
Triennial, AFSC.Slope, NWFSC.Combo, NWFSC.Slope, NWFSC.Shelf, NWFSC.Hypoxia,
NWFSC.Santa.Barb.Basin, NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working), NWFSC.Video}

\item{SaveFile}{option to save the file to the directory}

\item{Dir}{directory where the file should be saved}

\item{verbose}{opt to print out message statements}
}
\description{
Pull catch data from the [NWFSC](https://www.webapp.nwfsc.noaa.gov/data)
for a single species or all observed species, where the latter is specified
by leaving `Name = NULL` and `SciName = NULL`.
}
\examples{
\dontrun{
# SurveyName is only arg that has to be specified
dat <- PullCatch.fn(SurveyName = "NWFSC.Combo")

# Example with specified common name
catch_dat <- PullCatch.fn(Name = "vermilion rockfish",
SurveyName = "NWFSC.Combo")

# Example with specified scientific name
catch_dat <- PullCatch.fn(SciName = "Eopsetta jordani",
SurveyName = "NWFSC.Combo")

# Example with multiple names
catch_dat <- PullCatch.fn(Name = c("vermilion rockfish",
"vermilion and sunset rockfish"), SurveyName = "NWFSC.Combo")
catch_dat <- PullBio.fn(Name = c("vermilion rockfish",
"vermilion and sunset rockfish"), SurveyName = "NWFSC.Combo")
}

}
\author{
Chantel Wetzel based on code by John Wallace
}
