% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.fn.R
\name{PullCatch.fn}
\alias{PullCatch.fn}
\title{Pull catch data from the NWFSC data warehouse
The website is: https://www.nwfsc.noaa.gov/data
This function can be used to pull a single species or all observed species
In order to pull all species leave Name = NULL and SciName = NULL}
\usage{
PullCatch.fn(Name = NULL, SciName = NULL, YearRange = c(1000, 5000),
  SurveyName = NULL, SaveFile = FALSE, Dir = NULL, verbose = TRUE)
}
\arguments{
\item{Name}{common name of species data to pull from the data warehouse}

\item{SciName}{scientific name of species data to pull from the data warehouse}

\item{YearRange}{range of years to pull data}

\item{SurveyName}{survey to pull the data for the options are: Triennial, AFSC.Slope, NWFSC.Combo, NWFSC.Slope, NWFSC.Shelf, NWFSC.Hypoxia, NWFSC.Santa.Barb.Basin, NWFSC.Shelf.Rockfish, NWFSC.Video}

\item{SaveFile}{option to save the file to the directory}

\item{Dir}{directory where the file should be saved}

\item{verbose}{opt to print out message statements}
}
\description{
Pull catch data from the NWFSC data warehouse
The website is: https://www.nwfsc.noaa.gov/data
This function can be used to pull a single species or all observed species
In order to pull all species leave Name = NULL and SciName = NULL
}
\author{
Chantel Wetzel based on code by John Wallace
}
