% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.fn.R
\name{PullCatch.fn}
\alias{PullCatch.fn}
\title{Pull catch data from the NWFSC data warehouse
The website is: https://www.nwfsc.noaa.gov/data
Curently, this function only pulls data for a single specified survey.}
\usage{
PullCatch.fn(Species = "Sebastes pinniger", YearRange = c(1000, 5000),
  SurveyName = NULL, SaveFile = FALSE, Dir = NULL)
}
\arguments{
\item{Species}{scientific name of species data to pull from the data warehouse}

\item{YearRange}{range of years to pull data}

\item{SurveyName}{survey to pull the data for}

\item{SaveFile}{option to save the file to the directory}

\item{Dir}{directory where the file should be saved}
}
\description{
Pull catch data from the NWFSC data warehouse
The website is: https://www.nwfsc.noaa.gov/data
Curently, this function only pulls data for a single specified survey.
}
\author{
Chantel Wetzel based on code by John Wallac
}
