% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PullCatch.fn.R
\name{PullCatch.fn}
\alias{PullCatch.fn}
\title{Pull catch data from the NWFSC data warehouse}
\usage{
PullCatch.fn(
  Name = NULL,
  SciName = NULL,
  YearRange = c(1980, 5000),
  SurveyName = NULL,
  SaveFile = FALSE,
  Dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Name}{A character entry with the desired common name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{SciName} argument if you know the latin name.}

\item{SciName}{A character entry with the desired scientific name of the
species you want to pull data for from the data warehouse.
Use a vector of names if you want information for more than one species or
if the desired species is included in the database using more than one name,
e.g., vermilion rockfish (see the example below).
Use the \code{Name} argument if you know the common name.}

\item{YearRange}{An integer vector of length two with the
range of years to pull data for.}

\item{SurveyName}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf,
\item NWFSC.Hypoxia,
\item NWFSC.Santa.Barb.Basin,
\item NWFSC.Shelf.Rockfish (NWFSC.Hook.Line but both are not working),
\item NWFSC.Video.
}

Currently, you must pull data one survey at a time, though we are working on
allowing for a vector of survey names and
\code{NWFSC.Shelf.Rockfish} and \code{NWFSC.Hook.Line} are not supported.
The default of \code{NULL} is a placeholder that must be replaced with an entry.}

\item{SaveFile}{A logical value specifying whether or not the the data should
be saved to a file in \code{Dir}. Must change from the default of \code{FALSE} to save a file.}

\item{Dir}{If \code{SaveFile = TRUE}, then one must specify the directory where you want
the resulting file to be saved. The directory where the file should be saved.
The name of the file within \code{Dir} will start with Catch_ and end with .rda.}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Pull catch data from the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
for a single species or all observed species, where the latter is specified
by leaving both \code{Name = NULL} and \code{SciName = NULL}.
}
\details{
The data available in the warehouse are cleaned pior to being downloaded
with the intent that they provide the best available information for use
in an index-standardization procedure. The removed samples may be of use
to others with a less-restrictive goal than producing an index of abundance.
For example, life-stage samples are excluded because they are not collected
using the same protocols as standard samples.
To download all data, we currently recommend going to the
\href{https://www.webapps.nwfsc.noaa.gov/data}{NWFSC data warehouse}
and using the csv link to extract data for a single species at a time.
In the future, we hope to add functionality to this package such that
downloading all data can be done easily within this function.
See \href{https://github.com/pfmc-assessments/nwfscSurvey/issues/43}{Issue #43}
for more information.
}
\examples{
\dontrun{
# SurveyName is only arg that has to be specified
dat <- PullCatch.fn(SurveyName = "NWFSC.Combo")

# Example with specified common name
catch_dat <- PullCatch.fn(Name = "vermilion rockfish",
SurveyName = "NWFSC.Combo")

# Example with specified scientific name
catch_dat <- PullCatch.fn(SciName = "Eopsetta jordani",
SurveyName = "NWFSC.Combo")

# Example with multiple names
catch_dat <- PullBio.fn(Name = c("vermilion rockfish",
"vermilion and sunset rockfish"), SurveyName = "NWFSC.Combo")

}

}
\author{
Chantel Wetzel based on code by John Wallace
}
