% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInAges.fn.R
\name{ReadInAges.fn}
\alias{ReadInAges.fn}
\title{Reads in the West Coast Trienial survey data and filters the data into what is necessary
It reads in data and makes sure only the species necessary are kept
may want to keep NA (blank in Excel) to select the zero tows
removeCAN is a flag if you want tows in Canadian waters removed}
\usage{
ReadInAges.fn(dat, subset_years = NULL, verbose = TRUE)
}
\arguments{
\item{dat}{data file name}

\item{subset_years}{specify the years to retain, default is NULL which will provide 1977, alternative input would be 1980:2002 to remove only 1977.}

\item{verbose}{write out comments}
}
\description{
Necessary column names
   SPECIES_CODE
   AGE
}
\author{
Allan Hicks and Chantel Wetzel
}
