% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInExcelBiomass.fn.R
\name{readInExcelBiomass.fn}
\alias{readInExcelBiomass.fn}
\title{Reads in the stratum biomasses from the Excel file provided by Beth (BiomassAbundance Sheet)
headerRow is the row number of the column where the data start}
\usage{
readInExcelBiomass.fn(file, sheet = "BiomassAbundance", headerRow)
}
\arguments{
\item{file}{excel file name}

\item{sheet}{sheet name in excel file}

\item{headerRow}{line of header row in excel file}
}
\description{
Reads in the stratum biomasses from the Excel file provided by Beth (BiomassAbundance Sheet)
headerRow is the row number of the column where the data start
}
\author{
Allan Hicks & Chantel Wetzel
}
