% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInLengths.R
\name{ReadInLengths.fn}
\alias{ReadInLengths.fn}
\title{Reads in the compositional survey data and filters the data into what is necessary
It reads in catch and station data and makes sure only the species necessary are kept
### may want to keep NA (blank in Excel) to select the zero tows
removeCAN is a flag if you want tows in Canadian waters removed
   need the file called foreign_hauls.csv}
\usage{
ReadInLengths.fn(dat, verbose = TRUE)
}
\arguments{
\item{verbose}{print out comments}

\item{data}{data file name}
}
\description{
Necessary column names
   SPECIES_CODE
   LENGTH
}
\author{
Allan Hicks and Chantel Wetzel
}
