% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SexRatio.fn.R
\name{SexRatio.fn}
\alias{SexRatio.fn}
\title{Sex ratio function that assigns a sex to unsexed fish when creatin length or age comps}
\usage{
SexRatio.fn(
  x,
  sexRatioStage,
  sexRatioUnsexed,
  maxSizeUnsexed,
  bins = NULL,
  verbose
)
}
\arguments{
\item{x}{the data frame being passed to the function from either the SurveyLFs.fn or th SurveyAFs.fn}

\item{sexRatioStage}{the stage of the expansion to apply the sex ratio. Input either 1 or 2.}

\item{sexRatioUnsexed}{Sex ratio to apply to any length bins of a
certain size or smaller as defined by the maxSizeUnsexed. The default is \code{NA}.}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{bins}{Length bins created by the SurveyLFs.fn}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Sex ratio function that assigns a sex to unsexed fish when creatin length or age comps
}
\author{
Allan Hicks and Chantel Wetzel
}
