% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS3AF.fn.R
\name{SS3AF.fn}
\alias{SS3AF.fn}
\title{Wrapper function to create age frequency data in format for Stock Synthesis}
\usage{
SS3AF.fn(Age.df, ageBins, ...)
}
\arguments{
\item{Age.df}{age data frame}

\item{ageBins}{defined age bins}

\item{...}{additional inputs}
}
\description{
Wrapper function to create age frequency data in format for Stock Synthesis
}
\author{
Allan Hicks
}
