% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAFs.EWC.fn.R
\name{SurveyAFs.EWC.fn}
\alias{SurveyAFs.EWC.fn}
\title{Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datA should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SSout: if True the output is in a format pastable into SS dat file}
\usage{
SurveyAFs.EWC.fn(dir, datA, datTows, strat.vars = c("BOTTOM_DEPTH",
  "START_LATITUDE"), strat.df = NULL, femaleMale = c(2, 1), ageBins = 1,
  SSout = FALSE, meanRatioMethod = TRUE, gender = 3, NAs2zero = T,
  sexRatioUnsexed = NA, maxSizeUnsexed = NA, partition = 0,
  fleet = "Enter Fleet", agelow = "Enter", agehigh = "Enter",
  ageErr = "Enter", nSamps = "Enter Samps", month = "Enter Month",
  printfolder = "forSS")
}
\arguments{
\item{dir}{directory this is where the output files will be saved}

\item{datA}{the read in age data by the ReadInAges.EWC.fn function}

\item{datTows}{the read in catch data by the DesignBasedEstBiomass.EWC.fn function}

\item{strat.vars}{the variables used define the stratas. Defaul is bottom depth and latitudes.}

\item{strat.df}{the created strata matrix with the calculated areas by the createStrataDF.fn function}

\item{femaleMale}{numbering for female and male fish in the data file. This is opposite to what is used in SS.}

\item{SSout}{if True the output is in a format pastable into SS dat file}

\item{meanRatioMethod}{TRUE/FALSE}

\item{gender}{gender value for Stock Synthesis}

\item{NAs2zero}{change NAs to zeros}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{agelow}{value for SS -1}

\item{agehigh}{value for SS -1}

\item{ageErr}{age error vector to apply}

\item{nSamps}{effective sample size for Stock Synthesis}

\item{month}{month when the samples were collected}

\item{printfolder}{folder where the length comps will be saved}

\item{lageBins}{length bins}
}
\description{
Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datA should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SSout: if True the output is in a format pastable into SS dat file
}
\seealso{
\code{\link{SurveyLFs.EWC.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
