% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAFs.fn.R
\name{SurveyAFs.fn}
\alias{SurveyAFs.fn}
\title{Expands the ages up to the total stratum area then sums over strata
Original Version Written by Allan Hicks 16 March 2009
Modified by Chantel Wetzel to work with the data warehouse data formatting,
add additional options of when to apply the sex ratio, and correct some treatment of unsexed fish
weighted by sample size and area
NOTE: The age bin called F0 or M0 is retained to show proportion of ages smaller than smallest bin
You will want to likely add this to your first age bin and delete this before putting in SS, or
start the ageBins argument at the 2nd age bin and F0 will be all fish smaller (hence the first age bin)}
\usage{
SurveyAFs.fn(
  dir = NULL,
  datA,
  datTows,
  strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df = NULL,
  ageBins = 1,
  SSout = TRUE,
  meanRatioMethod = TRUE,
  sex = 3,
  NAs2zero = T,
  sexRatioUnsexed = NA,
  maxSizeUnsexed = NA,
  sexRatioStage = 1,
  partition = 0,
  fleet = "Enter Fleet",
  agelow = "Enter",
  agehigh = "Enter",
  ageErr = "Enter",
  nSamps = "Enter Samps",
  month = "Enter Month",
  printfolder = "forSS3",
  remove999 = TRUE,
  outputStage1 = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir}{directory this is where the output files will be saved}

\item{datA}{the biological data frame exctrated from the data warehouse using the PullBio.fn}

\item{datTows}{the catch data frame extracted from the data warehouse using the PullCatch.fn}

\item{strat.vars}{the variables used define the stratas. Defaul is bottom depth and latitudes.}

\item{strat.df}{the created strata matrix with the calculated areas by the createStrataDF.fn function}

\item{ageBins}{length bins}

\item{SSout}{if True the output is in a format pastable into SS dat file}

\item{meanRatioMethod}{TRUE/FALSE}

\item{sex}{sex value for Stock Synthesis}

\item{NAs2zero}{change NAs to zeros}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{sexRatioStage}{1/2 apply the sex ratio based on the tows (1) or the expanded numbers (2)}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{agelow}{value for SS -1}

\item{agehigh}{value for SS -1}

\item{ageErr}{age error vector to apply}

\item{nSamps}{effective sample size for Stock Synthesis}

\item{month}{month when the samples were collected}

\item{printfolder}{Folder inside the directory, if provide, where
csv files will be saved i.e., \code{file.path(dir, printfolder)}. Default "forSS3".}

\item{remove999}{the output object by the function will have the 999 column combined with the first length bin}

\item{outputStage1}{return the first stage expanded data without compiling it for SS}

\item{verbose}{A logical that specifies if you want to print messages and
warnings to the console. The default is \code{TRUE}.}
}
\description{
Expands the ages up to the total stratum area then sums over strata
Original Version Written by Allan Hicks 16 March 2009
Modified by Chantel Wetzel to work with the data warehouse data formatting,
add additional options of when to apply the sex ratio, and correct some treatment of unsexed fish
weighted by sample size and area
NOTE: The age bin called F0 or M0 is retained to show proportion of ages smaller than smallest bin
You will want to likely add this to your first age bin and delete this before putting in SS, or
start the ageBins argument at the 2nd age bin and F0 will be all fish smaller (hence the first age bin)
}
\seealso{
\code{\link{SurveyLFs.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
