% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyAgeAtLen.EWC.fn.R
\name{SurveyAgeAtLen.EWC.fn}
\alias{SurveyAgeAtLen.EWC.fn}
\title{Calculates proportion of age at length and reformats into SS3 format
Uses raw numbers at length, assuming that is a random sample conditioned on length and sex.
To use expanded numbers (up to strata areas), set raw=F
Only gender codes 1 and 2 and puts males and females on separate lines because the age@L is conditioned on sex (a sample of females of length 25cm, for example)
Gender=1: females only. Male values ignored
Gender=2: males only. Female values ignored.
lgthBins is either the interval between length bins or the actual length bins
note that 0 and Inf are tacked on the ends to account for lengths and ages outside the interval. You may want to add these in to first and last bin.
I assume all fish are sexed for age data, thus do not apply sex ratios for unsexed fish}
\usage{
SurveyAgeAtLen.EWC.fn(dir, datAL, datTows, strat.vars = NULL,
  strat.df = NULL, femaleMale = c(2, 1), lgthBins = 1, ageBins = 1,
  SS3out = F, meanRatioMethod = T, raw = TRUE, NAs2zero = T,
  season = "ENTER", fleet = "ENTER", partition = "ENTER",
  ageErr = "ENTER", returnSamps = F, printfolder = "forSS")
}
\arguments{
\item{dir}{directory location}

\item{datAL}{object}

\item{strat.df}{}

\item{lgthBins}{length bins}

\item{ageBins}{age bins}

\item{raw}{raw=T/F, input to define whether or not to expand numbers in the csv file (column header "NumF" and "NumM")}

\item{NAs2zero}{change NAs to zeros}

\item{season}{season}

\item{fleet}{fleet number}

\item{partition}{partition for Stock Synthesis}

\item{printfolder}{folder where the length comps will be saved}

\item{ageerr}{age error value for Stock Synthesis}
}
\description{
Calculates proportion of age at length and reformats into SS3 format
Uses raw numbers at length, assuming that is a random sample conditioned on length and sex.
To use expanded numbers (up to strata areas), set raw=F
Only gender codes 1 and 2 and puts males and females on separate lines because the age@L is conditioned on sex (a sample of females of length 25cm, for example)
Gender=1: females only. Male values ignored
Gender=2: males only. Female values ignored.
lgthBins is either the interval between length bins or the actual length bins
note that 0 and Inf are tacked on the ends to account for lengths and ages outside the interval. You may want to add these in to first and last bin.
I assume all fish are sexed for age data, thus do not apply sex ratios for unsexed fish
}
\seealso{
\code{\link{StrataFactors.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
