% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyLFs.fn.R
\name{SurveyLFs.fn}
\alias{SurveyLFs.fn}
\title{Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datL should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SSout: if True the output is in a format pastable into SS dat file}
\usage{
SurveyLFs.fn(dir, datL, datTows, strat.vars = c("Depth_m", "Latitude_dd"),
  strat.df = NULL, lgthBins = 1, SSout = TRUE, meanRatioMethod = TRUE,
  gender = 3, NAs2zero = T, sexRatioUnsexed = NA, maxSizeUnsexed = NA,
  sexRatioExpanded = FALSE, partition = 0, fleet = "Enter Fleet",
  nSamps = "Enter Samps", month = "Enter Month", printfolder = "forSS",
  remove999 = TRUE, switch = NULL, outputStage1 = FALSE)
}
\arguments{
\item{dir}{directory this is where the output files will be saved}

\item{datL}{the read in length comps by the ReadInLengths.EWC.fn function}

\item{datTows}{the read in catch data by the DesignBasedEstBiomass.EWC.fn function}

\item{strat.vars}{the variables used define the stratas. Defaul is bottom depth and latitudes.}

\item{strat.df}{the created strata matrix with the calculated areas by the createStrataDF.fn function}

\item{lgthBins}{length bins}

\item{SSout}{if True the output is in a format pastable into SS dat file}

\item{meanRatioMethod}{TRUE/FALSE}

\item{gender}{gender value for Stock Synthesis}

\item{NAs2zero}{change NAs to zeros}

\item{sexRatioUnsexed}{sex ratio to apply to any length bins of a certain size or smaller as defined by the maxSizeUnsexed}

\item{maxSizeUnsexed}{all sizes below this threshold will assign unsexed fish by sexRatio set equal to 0.50, fish larger than this size will have unsexed fish assigned by the calculated sex ratio in the data.}

\item{sexRatioExpanded}{switch to allow the user to apply the sex ratio post expanded numbers based on strata or raw data from each tow}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{nSamps}{effective sample size for Stock Synthesis}

\item{month}{month the samples were collected}

\item{printfolder}{folder where the length comps will be saved}

\item{remove999}{the output object by the function will have the 999 column combined with the first length bin}

\item{switch}{determined length or age data being processed}

\item{outputStage1}{return the first stage expanded data without compiling it for SS}
}
\description{
Epands the lengths up to the total stratum area then sums over strata
Written by Allan Hicks 16 March 2009
modified to incorporate unsexed fish using sex ratios in May 2011
weighted by sample size and area
datL should have a column called "year" indicating year
femaleMale is a vector of codes for female then male (in that order)
lgthBin is the increment of each length bin or a vector of the actual bins
NOTE: The length bin called F0 or M0 is retained to show proportion of lengths smaller than smallest bin
You will want to likely add this to your first length bin and delete this before putting in SS, or
start the lgthBins argument at the 2nd length bin and F0 will be all fish smaller (hence the first length bin)
SSout: if True the output is in a format pastable into SS dat file
}
\seealso{
\code{\link{StrataFactors.fn}}

\code{\link{SexRatio.fn}}
}
\author{
Allan Hicks and Chantel Wetzel
}
